% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccessConnect.R
\name{AccessConnect}
\alias{AccessConnect}
\title{Connect to an Access database}
\usage{
AccessConnect(path, silent = FALSE)
}
\arguments{
\item{path}{The path to the Access database file}

\item{silent}{TRUE suppresses messages except for errors.}
}
\value{
A connection to the database.
}
\description{
First checks that the proper driver exists, then attempts to connect to the database. A missing driver could trigger a crash of R, so it is important to check for it first. Can be used for all sorts of Access database connections, including EQWin.
}
