% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_location.R
\name{add_location}
\alias{add_location}
\title{Add location(s) to database}
\usage{
add_location(path, location, parameter, type)
}
\arguments{
\item{path}{The path to the local database, with extension.}

\item{location}{The location identifier, exactly as per the WSC or as written in Aquarius. Case sensitive. Specify as a character vector of length 1 or more.}

\item{parameter}{The parameter for this location, matched 1:1 to the locations vector. One of "level", "flow", "SWE", "snow depth", or "distance". The later is typically used for distances from bridge girders to water surface, but could be used for other distances.}

\item{type}{Continuous or discrete, matched to the locations and parameter vectors 1:1.}
}
\value{
One or more new entries are created in the table locations.
}
\description{
This function facilitates the addition of one or multiple stations to the database by adding entries to the locations table. The locations table and matching measurement tables will be populated during the next run of hydro_update_daily. If specifying level or flow, incorporation will first be attempted for a WSC location. Failure triggers an attempt to add a station from Aquarius.
}
