% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_vector.R
\name{add_vector}
\alias{add_vector}
\title{Add a vector file to the database}
\usage{
add_vector(con, table, geom)
}
\arguments{
\item{con}{A connection to the database, created with \code{\link[DBI:dbConnect]{DBI::dbConnect()}} or using the utility function \code{\link[=hydrometConnect]{hydrometConnect()}}.}

\item{table}{The referenced table in the database (as character string), must be under the public schema.}

\item{geom}{The geometry object to add to the database, as a \code{\link[terra:vect]{terra::vect()}} object. Conversion will automatically be made to epsg:4269, NAD83 lat/long decimal degrees.}
}
\value{
Nothing, but the vector should be added to the DB if no error is thrown.
}
\description{
Use this function to add a vector file such as a drainage polygon, provincial buffer, waterbodies, etc to the database. Ensures that database constraints are met. If you need to replace or delete a vector for any reason you'll have to use SQL (perhaps via R using the DBI package) to delete it first or using \code{\link[rpostgis:pgWriteGeom]{rpostgis::pgWriteGeom()}} with the necessary parameters.
}
