% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addACTimeseries.R
\name{addACTimeseries}
\alias{addACTimeseries}
\title{Add timeseries to aquacache database}
\usage{
addACTimeseries(
  df = NULL,
  data = NULL,
  start_datetime = NA,
  location = NA,
  sub_location = NA,
  z = NA,
  parameter = NA,
  media = NA,
  matrix_state_id = NA,
  sensor_priority = 1,
  aggregation_type = "instantaneous",
  record_rate = NA,
  share_with = "public_reader",
  owner = NA,
  source_adapters = NULL,
  note = NA,
  con = NULL
)
}
\arguments{
\item{df}{A data.frame containing at least one row and the following columns: start_datetime, location, z, parameter, media, sensor_priority, aggregation_type, record_rate, share_with, owner, note. An optional list-column named \code{source_adapters} may contain one assignment data.frame per new timeseries. If this parameter is provided, all other parameters save for \code{data} must be left at their defaults.}

\item{data}{An optional list of data.frames of length nrow(df) or length(location) containing the data to add to the database. If adding multiple timeseries and not all of them need data, include NA elements in the list in the correct locations.}

\item{start_datetime}{A character or POSIXct vector of datetimes from which
to look for new data when an active fetch assignment is specified. Values
are coerced to POSIXct in UTC.}

\item{location}{A character vector corresponding to locations.location_code (preferred), locations.alias (legacy and nullable), or locations.name OR a numeric vector corresponding to locations.location_id.}

\item{sub_location}{A numeric vector corresponding to column 'sub_location_id' of table 'sub_locations'. This is optional and can be left as NA if not specified. It is used to differentiate between multiple timeseries at the same location, e.g. different standpipes or wells.}

\item{z}{A numeric vector of elevations in meters for the timeseries observations. This allows for differentiation of things like wind speeds at different heights. Leave as NA if not specified.}

\item{parameter}{A numeric vector corresponding to column 'parameter_id' of table 'parameters'.}

\item{media}{A numeric vector corresponding to column 'media_id' of table 'media_types'.}

\item{matrix_state_id}{An optional numeric vector corresponding to column 'matrix_state_id' of table 'matrix_states'. Leave as NA to let the database resolve the matrix state from the media and parameter defaults.}

\item{sensor_priority}{A numeric vector assigning priority order to assign to this timeseries, default 1. This can allow for storage of multiple identical timeseries taken by different sensors for redundancy.}

\item{aggregation_type}{A character vector describing the measurement type; one of 'instantaneous' (immediate sensor value), 'sum', 'mean', 'median', 'min', 'max', '(min+max)/2'.}

\item{record_rate}{A broad categorization of the rate at which recording takes place. Select from a number fo minutes or hours ('5 minutes', '1 hour'), '1 day', '1 week', '4 weeks', '1 month', '1 year'.}

\item{share_with}{A \emph{character} vector of the user group(s) with which to share the timeseries, Default is 'public_reader'. Pass multiple groups as a single string, e.g. "public_reader, YG" or multiple such strings if specifying multiple timeseries in one go.}

\item{owner}{A numeric vector of the owner(s) of the timeseries(s). This can be different from the location owner!}

\item{source_adapters}{\code{NULL}, one assignment data.frame when adding one
timeseries, or a list containing one assignment data.frame (or \code{NULL}) per
new timeseries. Columns are \code{source_fx}, optional JSON or named-list
\code{source_fx_args}, \code{fetch_priority}, \code{synchronize_priority}, \code{active}, and
\code{note}. Every function must be registered for the continuous domain.}

\item{note}{Text notes to append to the timeseries.}

\item{con}{A connection to the database, created with \code{\link[DBI:dbConnect]{DBI::dbConnect()}} or using the utility function \code{\link[=AquaConnect]{AquaConnect()}}. Leave NULL to use the package default connection and have it closed afterwards automatically.}
}
\value{
One or more new entries are created in the table 'timeseries'
}
\description{
This function facilitates the addition of one or multiple timeseries to the database by adding entries to the timeseries and settings tables. See related function \code{\link[=addACLocation]{addACLocation()}} for adding a location to which timeseries must be attached. To add an image series see \code{\link[=addACImageSeries]{addACImageSeries()}}, for raster series see \code{\link[=addACRasterSeries]{addACRasterSeries()}}. For one-off images use \code{\link[=insertACImage]{insertACImage()}} and for rasters \code{\link[=insertACRaster]{insertACRaster()}}. For documents use \code{\link[=insertACDocument]{insertACDocument()}}.

You will be prompted to add locations that don't exist yet if any fall into this category.
}
\details{
You can add the new timeseries by directly editing the database, but this function ensures that database constraints are respected and will immediately seek to populate the measurements and calculated tables with new information for each timeseries. For Water Survey of Canada data, this function will also seek out level and flow data from the HYDAT database, downloading or checking for updates to it before use.

If specifying a data.frame for argument \code{data}, different criteria applies depending on if the timeseries is categorized as continuous or discrete.
For continuous data:
The data.frame must contain a 'datetime' (POSIXct) OR 'date' (date) column. If specifying 'date' then the data is entered to \code{measurements_continuous} as one-day-period rows and database triggers maintain \code{measurements_calculated_daily}. 'value' (numeric) is also required, and optionally 'owner', 'contributor', 'share_with', 'approval', 'grade', 'qualifier'. Function \code{\link[=addNewContinuous]{addNewContinuous()}} will be called to add this data to the database. If a fetch source adapter is also specified it will be called to fetch more recent data than that in this data.frame.

Source adapters are supplied in \code{source_adapters}. Each assignment may be used for fetching, synchronization, or both, and may be retained while temporarily inactive. The highest active priority (lowest number) is selected for each operation.
}
\examples{
\dontrun{
# Each timeseries receives its own source-adapter assignment data frame.
precipitation_sources <- data.frame(
  source_fx = c("downloadAquarius", "downloadRWIS"),
  source_fx_args = I(list(
    list(
      location = "09AA-M3",
      parameter = "Precip Total",
      difference = TRUE
    ),
    list(location = "RWIS_STATION", parameter = "Precipitation")
  )),
  fetch_priority = c(1L, 2L),
  synchronize_priority = c(2L, 1L),
  active = c(TRUE, TRUE),
  note = c(
    "Preferred source for routine fetching.",
    "Preferred source for full synchronization."
  )
)

wind_sources <- data.frame(
  source_fx = "downloadAquarius",
  source_fx_args = I(list(list(
    location = "09AA-M3",
    parameter = "Wind Speed"
  ))),
  fetch_priority = 1L,
  synchronize_priority = 1L,
  active = TRUE
)

# Add the assignment data frames as a list-column, one per timeseries.
df <- data.frame(
  start_datetime = "2015-01-01 00:00",
  location = "09AA-M3",
  z = c(NA, 3),
  parameter = c(34, 1154),
  media = 7,
  sensor_priority = 1,
  aggregation_type = c("sum", "mean"),
  record_rate = "1 hour",
  share_with = "public_reader",
  owner = 2,
  note = c(
    "Total precipitation from standpipe, reset every fall.",
    "Hourly average of wind speeds recorded every minute."
  )
)
df$source_adapters <- list(precipitation_sources, wind_sources)

addACTimeseries(df)
}
}
