% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addACUser.R
\name{addACUser}
\alias{addACUser}
\title{Create a new user in the aquacache database}
\usage{
addACUser(
  username,
  password,
  email,
  group_ids = c(1),
  con = AquaConnect(silent = TRUE),
  algorithm = "sha256"
)
}
\arguments{
\item{username}{The username for the new user.}

\item{password}{The password for the new user.}

\item{email}{The email address for the new user.}

\item{group_ids}{The group_ids for the new user. Default 1 is the 'public' group. To add a group, use \code{\link[=addACUserGroup]{addACUserGroup()}}.}

\item{con}{A connection to the aquacache database. Default uses \code{\link[=AquaConnect]{AquaConnect()}}.}

\item{algorithm}{The hashing algorithm to use. Default is "sha256". Refer to \code{\link[digest:digest]{digest::digest()}} for available algorithms.}
}
\value{
The new user is added to the database.
}
\description{
The aquacache database uses row-level security to restrict record access to users based on their group_id.
Users are stored in the 'users' table in the aquacache database. The password is hashed using the specified algorithm and a random salt.
}
