% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addACUserGroup.R
\name{addACUserGroup}
\alias{addACUserGroup}
\title{Add a new user group to the AquaCache}
\usage{
addACUserGroup(name, description, con = AquaConnect(silent = TRUE))
}
\arguments{
\item{name}{The name of the new user group}

\item{description}{A description of the new user group}

\item{con}{A connection to the AquaCache database. Default uses \code{\link[=AquaConnect]{AquaConnect()}}.}
}
\value{
The group_id of the new user group, plus a new entry to the database.
}
\description{
Add a new user group to the AquaCache
}
