% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addHydrometDocument.R
\name{addHydrometDocument}
\alias{addHydrometDocument}
\title{Add document to hydrometric database}
\usage{
addHydrometDocument(
  path,
  location,
  description,
  con = hydrometConnect(silent = TRUE)
)
}
\arguments{
\item{path}{Valid path including extension to the document to upload.}

\item{location}{The location with which to associate the document (must be in the database).}

\item{description}{A text description of what the document is. Please be descriptive!}

\item{con}{A connection to the database.}
}
\value{
TRUE if a document was properly added to the database.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function facilitates the addition of one document at a time to the database in the 'documents' table. Each document must be linked to a specific location. Adding a document directly to the database is not possible, since the file must be converted to a binary object before loading. See \code{\link[=fetchDocument]{fetchDocument()}} to get a document out again.
}
