% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addHydrometImages.R
\name{addHydrometImages}
\alias{addHydrometImages}
\title{Add images to hydrometric database}
\usage{
addHydrometImages(
  path,
  location,
  image_type,
  datetime,
  fetch_datetime = NULL,
  con = hydrometConnect(silent = TRUE)
)
}
\arguments{
\item{path}{Valid path including extension to the document to upload.}

\item{location}{The location with which to associate the document (must be in the database).}

\item{image_type}{The type of image: 'auto', or 'manual'.}

\item{datetime}{The datetime the image was taken at, as a POSIXct object or something that can be coerced to one. If not POSIXct timezone is assumed to be UTC.}

\item{fetch_datetime}{The datetime the image was retrieved (optional).}

\item{con}{A connection to the database.}
}
\value{
TRUE if an image was properly added to the database.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function facilitates the addition of one image at a time to the database in the 'images' table. Each image must be linked to a specific location. Adding an image directly to the database is not possible, since the file must be converted to a binary object before loading. See \code{\link[=fetchDocument]{fetchDocument()}} to get the image out again.
}
