% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addHydrometImageSeries.R
\name{addHydrometImageSeries}
\alias{addHydrometImageSeries}
\title{Add new image series (auto-fetched)}
\usage{
addHydrometImageSeries(
  location,
  start_datetime,
  source_fx,
  source_fx_args = NA,
  public = TRUE,
  public_delay = NA,
  con = hydrometConnect(silent = TRUE)
)
}
\arguments{
\item{location}{The location associated with the image series.}

\item{start_datetime}{The datetime (as POSIXct) from which to look for images}

\item{source_fx}{The function to use for fetching new images. Must be an existing function in this package.}

\item{source_fx_args}{Additional arguments to pass to the function, in the form "{param1 = arg1}, {param2 = 'arg2'}". Do not deviate from this format!}

\item{public}{Should the function be publicly visible?}

\item{public_delay}{A period in ISO 8601 format by which to delay public visibility of images.}

\item{con}{A connection to the database, created with \code{\link[DBI:dbConnect]{DBI::dbConnect()}} or using the utility function \code{\link[=hydrometConnect]{hydrometConnect()}}.}
}
\value{
TRUE if successful, and a new entry in the database with images fetched.
}
\description{
Add new image series (auto-fetched)
}
