% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addHydrometRasterSeries.R
\name{addHydrometRasterSeries}
\alias{addHydrometRasterSeries}
\title{Add new raster series (auto-fetched models)}
\usage{
addHydrometRasterSeries(
  model,
  start_datetime,
  source_fx,
  source_fx_args = NA,
  public = TRUE,
  public_delay = NA,
  con = hydrometConnect(silent = TRUE)
)
}
\arguments{
\item{model}{The model producing the raster}

\item{start_datetime}{The datetime (as POSIXct) from which to look for rasters}

\item{source_fx}{The function to use for fetching new rasters Must be an existing function in this package.}

\item{source_fx_args}{Additional arguments to pass to the function, in the form "\{param1 = arg1\}, \{param2 = 'arg2'\}". Each parameter = value pair needs to be enclosed in curly brackets, which might be missing here. Do not deviate from this format!}

\item{public}{Should the rasters be publicly visible?}

\item{public_delay}{A period in ISO 8601 format by which to delay public visibility of images.}

\item{con}{A connection to the database, created with \code{\link[DBI:dbConnect]{DBI::dbConnect()}} or using the utility function \code{\link[=hydrometConnect]{hydrometConnect()}}.}
}
\value{
TRUE if successful, and a new entry in the database with images fetched.
}
\description{
Use this function to add a new series of rasters generated by a model, such as the RDPS or HRDPS models by ECCC.
}
