% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addHydrometTemplate.R
\name{addHydrometTemplate}
\alias{addHydrometTemplate}
\title{Templates to add timeseries to hydrometric database}
\usage{
addHydrometTemplate(
  con = hydrometConnect(silent = TRUE),
  format = "short",
  save_path = NULL
)
}
\arguments{
\item{con}{A connection to the database, created with \code{\link[DBI:dbConnect]{DBI::dbConnect()}} or using the utility function \code{\link[=hydrometConnect]{hydrometConnect()}}.}

\item{format}{"long" or "short", only applies to the timeseries table. If long, will return one row per unique parameter with the total data.frame length equal to the parameter with the highest count. If short will return a single row, with unique parameter values concatenated in a single cell per parameter.}

\item{save_path}{Specify a save path (folder) if you want an xlsx document with column headers as required by function \code{\link[=addHydrometTimeseries]{addHydrometTimeseries()}}. "choose" lets you choose interactively.}
}
\value{
A list of three data.frames:
For the timeseries table, a data.frame with column names and options already in the DB for each required parameter to pass to the parameter \code{timeseries_df} of \code{\link[=addHydrometTimeseries]{addHydrometTimeseries()}}.
For the spatial table, a data.frame with column names and options already in the DB for each required parameter to pass to the parameter \code{locations_df} of \code{\link[=addHydrometTimeseries]{addHydrometTimeseries()}}.
A data.frame is also produced listing the datums present in the database, facilitating your specification of the appropriate datum in the spatial table.
}
\description{
Additional arguments to pass to the function specified in source_fx should take the form of "{param1 = arg1}, {param2 = 'arg2'}". The data fetch function will separate out the parameter:argument pairs based on them being within curly brackets.
}
