% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addHydrometTimeseries.R
\name{addHydrometTimeseries}
\alias{addHydrometTimeseries}
\title{Add timeseries to hydrometric database}
\usage{
addHydrometTimeseries(
  con = hydrometConnect(silent = TRUE),
  timeseries_df,
  locations_df = NULL
)
}
\arguments{
\item{con}{A connection to the database, created with \code{\link[DBI:dbConnect]{DBI::dbConnect()}} or using the utility function \code{\link[=hydrometConnect]{hydrometConnect()}}.}

\item{timeseries_df}{A data.frame containing the information necessary to add the timeseries (see details for template).}

\item{locations_df}{A data.frame containing spatial information related to the individual locations specified in timeseries_df. Only necessary if you are specifying a location code that is NOT already in the database. Function returns an error if you didn't specify a spatial_df when it is necessary. See details for template.}
}
\value{
One or more new entries are created in the table 'timeseries'
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function facilitates the addition of one or multiple timeseries to the database by adding entries to the timeseries, locations, and datum_conversions tables. See related function \code{\link[=addHydrometTemplate]{addHydrometTemplate()}} for help in formatting the data.frames to pass to \code{timeseries_df} and \code{locations_df}.
}
\details{
You can also add the new timeseries by directly editing the database, but this function ensures that database constraints are respected and will immediately seek to populate the measurements and calculated tables with new information for each timeseries.

Additional arguments to pass to the function specified in source_fx should take the form of "{param1 = arg1}, {param2 = 'arg2'}". The data fetch function will separate out the parameter:argument pairs based on them being within curly brackets.
}
\seealso{
\code{\link[=addHydrometTemplate]{addHydrometTemplate()}} to see templates for timesries_df and locations_df.
}
