% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{advisory_lock_acquire}
\alias{advisory_lock_acquire}
\title{Acquire a PostgreSQL advisory lock for a namespaced key.}
\usage{
advisory_lock_acquire(con, namespace, key, wait = TRUE)
}
\arguments{
\item{con}{A database connection object.}

\item{namespace}{The namespace for the lock.}

\item{key}{The key for the lock.}

\item{wait}{A boolean indicating whether to wait for the lock (default: TRUE).}
}
\value{
A boolean indicating whether the lock was acquired.
}
\description{
Attempts to acquire a PostgreSQL advisory lock for a given namespace and key. If \code{wait} is TRUE, will block until the lock is acquired. If \code{wait} is FALSE, will return immediately with a boolean indicating whether the lock was acquired.
}
