% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{advisory_lock_release}
\alias{advisory_lock_release}
\title{Release a PostgreSQL advisory lock for a namespaced key.}
\usage{
advisory_lock_release(con, namespace, key)
}
\arguments{
\item{con}{A database connection object.}

\item{namespace}{The namespace for the lock.}

\item{key}{The key for the lock.}
}
\value{
A boolean indicating whether the lock was released.
}
\description{
Includes a recovery path for aborted transactions and a final fallback to unlock all session advisory locks.
}
