% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AquaCache-package.R
\docType{package}
\name{AquaCache-package}
\alias{AquaCache}
\alias{AquaCache-package}
\title{AquaCache: Hydrometeorological database creation and maintenance}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Set of functions for the ongoing maintenance of the Yukon Government's hydrometeorological databases in postgreSQL format. Contains functions which operate on two databases, the 'snow' and 'aquacache' databases. The former is designed to hold discrete snowpack measurements and detailled snow survey data, while the later is designed to hold hydrometric, meteorological and water quality data along with images, spatial files and documents. Currently includes adapters to import data from the Water Survey of Canada (WSC) real-time and historical data, data held in Aquarius, USGS data, discrete snow survey data from the snow database created with this package, water quality data, images from NuPoint and the WSC, and meteorological data in the form of timeseries observations at locations and rasters data (with special handling for forecast rasters). Supports any number of timeseries unique by their location, parameter, data type, and period (for sum, mean, etc. measurements). Disparate data sources are handled via import scripts and function arguments specified for each timeseries/sample series/raster series/image series, allowing for great flexibility.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/YukonWRB/AquaCache}
  \item Report bugs at \url{https://github.com/YukonWRB/AquaCache/issues}
}

}
\author{
\strong{Maintainer}: Ghislain de Laplante \email{ghislain.delaplante@yukon.ca} (\href{https://orcid.org/0000-0002-5093-9185}{ORCID})

Authors:
\itemize{
  \item Ghislain de Laplante \email{ghislain.delaplante@yukon.ca} (\href{https://orcid.org/0000-0002-5093-9185}{ORCID})
}

Other contributors:
\itemize{
  \item Emilie Stewart-Jones \email{emilie.stewart-jones@yukon.ca} (\href{https://orcid.org/0000-0002-0013-8953}{ORCID}) [contributor]
  \item Everett Snieder \email{everett.snieder@gmail.com} (\href{https://orcid.org/0000-0003-4997-3404}{ORCID}) [contributor]
}

}
\keyword{internal}
