% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AquaCacheApp.R
\name{AquaCacheApp}
\alias{AquaCacheApp}
\title{Shiny Interface to Hydromet DB}
\usage{
AquaCacheApp()
}
\value{
Opens the Shiny application.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Loads a Shiny app that facilitates the addition of information to the database.
Currently supports uploading of documents, one-off images, vectors, and raster, with provisions to support adding timeseries (data, images, and rasters).

Requires write privileges to the database.
}
