% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AquaCacheConnect.R
\name{AquaCacheCon}
\alias{AquaCacheCon}
\title{Connect to the AquaCache database}
\usage{
AquaCacheCon(
  name = "AquaCache",
  host = Sys.getenv("AquaCacheHost"),
  port = Sys.getenv("AquaCachePort"),
  username = Sys.getenv("AquaCacheAdminUser"),
  password = Sys.getenv("AquaCacheAdminPass"),
  silent = FALSE
)
}
\arguments{
\item{name}{Database name.}

\item{host}{Database host address. By default searches the .Renviron file for parameter:value pair of form AquaCacheHost:"hostname".}

\item{port}{Connection port. By default searches the .Renviron file for parameter:value pair of form AquaCachePort:"1234".}

\item{username}{Username. By default searches the .Renviron file for parameter:value pair of form AquaCacheAdminUser:"username". Refrain from using username with write privileges unless you absolutely know what you're doing.}

\item{password}{Password. By default searches the .Renviron file for parameter:value pair of form AquaCacheAdminPass:"password".}

\item{silent}{TRUE suppresses messages except for errors.}
}
\value{
A connection to the database.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function exists to facilitate connecting to the hydrology database, especially if the database type and connection method changes in the future.
}
