% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AquaCacheInit.R
\name{AquaCacheInit}
\alias{AquaCacheInit}
\title{Initial hydrometric/meteorological database creation.}
\usage{
AquaCacheInit(con = AquaConnect(), overwrite = FALSE)
}
\arguments{
\item{con}{A connection to the database, created with \code{\link[DBI:dbConnect]{DBI::dbConnect()}} or using the utility function \code{\link[=AquaConnect]{AquaConnect()}}.}

\item{overwrite}{TRUE overwrites the database, if one exists. Nothing will be kept. FALSE will create tables only where they are missing.}
}
\value{
New tables in the target postgres database.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Populates a postgreSQL database or replaces existing tables. Establishes pre-set schemas and populates some tables with initial values. No indices are specified as the primary keys fulfill this task already.
}
