% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AquaPatchCheck.R
\name{AquaPatchCheck}
\alias{AquaPatchCheck}
\title{Check for patches to apply to the AquaCache database}
\usage{
AquaPatchCheck(con)
}
\arguments{
\item{con}{A connection to the database, created with \code{\link[DBI:dbConnect]{DBI::dbConnect()}} or using the utility function \code{\link[=AquaConnect]{AquaConnect()}}.}
}
\value{
A message indicating whether or not patches need to be applied.
}
\description{
The functions in this package are dependent on a consistent database schema. However, sometimes changes need to be made to the database schema either for the sake of data organization in the DB or to facilitate or allow new functionality via this package. This function checks for patches that need to be applied to the database to ensure that the schema is up-to-date before running any functions that depend on it.
}
