% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dailyCrossCheck.R
\name{dailyCrossCheck}
\alias{dailyCrossCheck}
\title{Cross-check daily and realtime tables for consistency}
\usage{
dailyCrossCheck(path, locations = "all")
}
\arguments{
\item{path}{The path to the local hydro SQLite database, with extension. Passed to \code{\link[WRBtools:hydroConnect]{WRBtools::hydroConnect()}}.}

\item{locations}{The locations you wish to have updated as a character vector. Defaults to "all", though the meaning of all is dependent on the parameter 'aquarius'. Will search for all timeseries with the specified location codes, so level + flow, snow depth + SWE, etc.}
}
\value{
Updated entries in the hydro database.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Checks that there is a corresponding entry in table "daily" for every day in table "realtime" for every location specified in \code{locations}.
}
