% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dailyUpdate.R
\name{dailyUpdate}
\alias{dailyUpdate}
\title{Daily update of hydro database}
\usage{
dailyUpdate(
  con = NULL,
  timeseries_id = "all",
  sample_series_id = "all",
  active = "default",
  continuous = TRUE,
  discrete = TRUE,
  hydat = TRUE,
  images = TRUE,
  rasters = TRUE
)
}
\arguments{
\item{con}{A connection to the database, created with \code{\link[DBI:dbConnect]{DBI::dbConnect()}} or using the utility function \code{\link[=AquaConnect]{AquaConnect()}}. NULL will create a connection and close it afterwards, otherwise it's up to you to close it after.}

\item{timeseries_id}{The timeseries_ids you wish to have updated, as character or numeric vector. Defaults to "all".}

\item{sample_series_id}{The sample_series_ids you wish to have updated, as character or numeric vector. Defaults to "all".}

\item{active}{Sets behavior for import of new data. If set to 'default', the function will look to the column 'active' in the 'timeseries', 'image_series', or 'raster_series_index' tables to determine if new data should be fetched. If set to 'all', the function will ignore the 'active' column and import all data.}

\item{continuous}{If TRUE, will update continuous data. Default is TRUE.}

\item{discrete}{If TRUE, will update discrete data. Default is TRUE.}

\item{hydat}{If TRUE, will check for new HYDAT data and update continuous.timeseries in the database if needed. Default is TRUE.}

\item{images}{If TRUE, will fetch new images. Default is TRUE.}

\item{rasters}{If TRUE, will fetch new rasters. Default is TRUE.}
}
\value{
The database is updated in-place, and diagnostic messages are printed to the console.
}
\description{
This function is intended to be run on a daily or near-daily basis to ensure data integrity. Pulls in new data, calculates statistics where necessary, and performs cross-checks on several tables (see details for more information).
}
\details{
Calls several functions in sequence: \code{\link[=getNewContinuous]{getNewContinuous()}} to pull in new data into the measurements_continuous table and update daily statistics for those imports, \code{\link[=getNewDiscrete]{getNewDiscrete()}} to pull in new data to the measurements_discrete table, \code{\link[=getNewImages]{getNewImages()}} to get new images in a series, \code{\link[=getNewRasters]{getNewRasters()}} to get new model rasters, \code{\link[=update_hydat]{update_hydat()}} to check for a new HYDAT database version (hydrometric data from the WSC) and incorporate daily means which differ from those already calculated in the dabatabase, and \code{\link[=update_hydat_datums]{update_hydat_datums()}} to update the datums table with any new datums present in HYDAT.

Note that new timeseries should be added using function \code{\link[=addACTimeseries]{addACTimeseries()}}.

Continuous and discrete series participate when they have an active fetch
assignment. A selected \code{downloadAquarius} assignment needs credentials in
\code{.Renviron} or in its \code{source_fx_args}; see \code{\link[=downloadAquarius]{downloadAquarius()}}.
}
