% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_dump.R
\name{db_dump}
\alias{db_dump}
\title{Dump database schema and data to a file}
\usage{
db_dump(
  name = "aquacache",
  host = Sys.getenv("aquacacheHost"),
  port = Sys.getenv("aquacachePort"),
  username = Sys.getenv("aquacacheAdminUser"),
  password = Sys.getenv("aquacacheAdminPass"),
  outpath = "choose",
  pg_dump = NULL,
  pg_dumpall = NULL,
  data = FALSE
)
}
\arguments{
\item{name}{Database name.}

\item{host}{Database host address. By default searches the .Renviron file for parameter:value pair of form aquacacheHost="hostname".}

\item{port}{Connection port. By default searches the .Renviron file for parameter:value pair of form aquacachePort="1234".}

\item{username}{Username. By default searches the .Renviron file for parameter:value pair of form aquacacheAdminUser="username".}

\item{password}{Password. By default searches the .Renviron file for parameter:value pair of form aquacacheAdminPass="password".}

\item{outpath}{The path to the folder/location where the schema dump will be saved. If "choose", the user will be prompted to select a folder. Separate files will be created for the outputs of pg_dumpall and pg_dump.}

\item{pg_dump}{The path to the pg_dump utility. By default (NULL), the function searches the PATH for the utility, but this might not always work.}

\item{pg_dumpall}{The path to the pg_dumpall utility. By default (NULL), the function searches the PATH for the utility, but this might not always work.}

\item{data}{Should the contents of the target database be included in the dump? Default is FALSE.}
}
\value{
An SQL file containing the schema definition saved to the 'outfile' path.
}
\description{
This function uses the \code{pg_dump} and \code{pg_dumpall} utilities to create a dump of the schema and global objects in a PostgreSQL database. The schema dump is saved to an SQL file in the specified output path. You must call this function from a machine with the PostgreSQL utilities installed.
}
\seealso{
\code{\link[=create_test_db]{create_test_db()}}
}
