% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dbTransBegin}
\alias{dbTransBegin}
\title{Begin a transaction}
\usage{
dbTransBegin(con, silent = TRUE)
}
\arguments{
\item{con}{A database connection object.}

\item{silent}{A boolean indicating whether to suppress messages about transaction status.}
}
\value{
A boolean indicating whether a transaction was started
}
\description{
Begin a transaction on a database connection. This function creates a temporary table to ensure the transaction is active, which DBI::dbBegin does not do. The transaction must be committed or rolled back using \code{DBI::dbExecute(con, "COMMIT;")} or \code{DBI::dbExecute(con, "ROLLBACK;")}. Note that you should not use DBI's built-in transaction functions to commit or rollback a transaction started with this function.
}
