% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dbTransCheck}
\alias{dbTransCheck}
\title{Check if a transaction is active}
\usage{
dbTransCheck(con)
}
\arguments{
\item{con}{A database connection object.}
}
\value{
A boolean indicating whether a transaction is active (TRUE for active)
}
\description{
Check if a database connection is in an active transaction.
}
