% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadCaLDAS.R
\name{downloadCaLDAS}
\alias{downloadCaLDAS}
\title{Get CaLDAS rasters}
\usage{
downloadCaLDAS(parameter, start_datetime, clip = NULL, hrs = c(0))
}
\arguments{
\item{parameter}{The parameter for which to get new rasters, such as 'AirTemp_AGL-1.5m' or 'SnowDepth_Sfc' or 'SnowWaterEquiv_Sfc'.}

\item{start_datetime}{The datetime from which to start looking for new rasters. Coerced to POSIXct, timezone UTC. At present, CaLDAS rasters are only available for 30 days in the past, so earlier start datetimes will be adjusted to 30 days ago for speed.}

\item{clip}{The area with which to clip the rasters. Supply one or more
two-letter province or territory abbreviations as per the \href{https://www12.statcan.gc.ca/census-recensement/2021/ref/dict/tab/index-eng.cfm?ID=t1_8}{Canadian Census}
to use their combined 300 km buffered boundary. Alternatively, supply an
unnamed numeric EPSG:4326 vector in \code{north}, \code{west}, \code{south}, \code{east} order, a named
numeric vector or named list using either those names or \code{xmin}, \code{xmax},
\code{ymin}, and \code{ymax}, or a \link[terra:SpatExtent]{terra::SpatExtent}. JSON arrays and named JSON
objects passed through \code{source_fx_args} are supported. Set to \code{NULL} for
no clip.}

\item{hrs}{CaLDAS data is provided every 3 hours. Give a vector of hours from 0, 3, 6, 9, 12, 15, 18, 21 to specify which rasters to bring in from start_datetime to end_datetime. Default is for 0 hours only, so each day at 00:00 UTC.}
}
\value{
A list of lists, where each element consists of the target raster as well as associated attributes.
}
\description{
Get CaLDAS rasters
}
