% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadHRDPS.R
\name{downloadHRDPS}
\alias{downloadHRDPS}
\title{Get HRDPS rasters}
\usage{
downloadHRDPS(parameter, start_datetime, clip = NULL)
}
\arguments{
\item{parameter}{The parameter to fetch.}

\item{start_datetime}{Not a true start_datetime in the sense used in fetching other data, but rather the datetime of the last issued forecast in the database. This is compared to what's on the remote and, if different, the new forecast is fetched.}

\item{clip}{The area with which to clip the rasters. Supply one or more
two-letter province or territory abbreviations as per the \href{https://www12.statcan.gc.ca/census-recensement/2021/ref/dict/tab/index-eng.cfm?ID=t1_8}{Canadian Census}
to use their combined 300 km buffered boundary. Alternatively, supply an
unnamed numeric EPSG:4326 vector in \code{north}, \code{west}, \code{south}, \code{east} order, a named
numeric vector or named list using either those names or \code{xmin}, \code{xmax},
\code{ymin}, and \code{ymax}, or a \link[terra:SpatExtent]{terra::SpatExtent}. JSON arrays and named JSON
objects passed through \code{source_fx_args} are supported. Set to \code{NULL} for
no clip.}
}
\value{
A list of lists, where each element consists of the target raster as well as associated attributes.
}
\description{
Get HRDPS rasters
}
