% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadNupointImages.R
\name{downloadNupointImages}
\alias{downloadNupointImages}
\title{Get images from NuPoint SFTP server}
\usage{
downloadNupointImages(
  location,
  start_datetime,
  username = Sys.getenv("nupointUser"),
  password = Sys.getenv("nupointPass"),
  url = Sys.getenv("nupointServer"),
  port = Sys.getenv("nupointPort"),
  folder = Sys.getenv("nupointFolder"),
  save_path = NULL,
  delete = TRUE
)
}
\arguments{
\item{location}{The location for which to get images.}

\item{start_datetime}{The earliest datetime to start pulling images from.}

\item{username}{Username to use for password-protected login.}

\item{password}{Password to use for protected login.}

\item{url}{The URL or IP address from which to get new images.}

\item{port}{The port on which to connect to the SFTP server.}

\item{folder}{The folder in which to look for new images.}

\item{save_path}{Optional; path in which to save the image.}

\item{delete}{Should the files be deleted from the SFTP site once they've
been successfully fetched?}
}
\value{
A data.frame containing downloaded file paths, datetimes, and
locations, or NULL if no matching files are found.
}
\description{
Fetches images from a NuPoint SFTP server. To fetch from the WSC instead see
\code{\link[=downloadWSCImages]{downloadWSCImages()}}. Intended to be called by function \code{\link[=getNewImages]{getNewImages()}}.
Use caution: this function will delete files from the SFTP server after
fetching if \code{delete = TRUE} (the default).
}
