% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadNWIS.R
\name{downloadNWIS}
\alias{downloadNWIS}
\title{Get realtime data from NWIS locations}
\usage{
downloadNWIS(
  location,
  parameter_id,
  start_datetime,
  end_datetime = Sys.time(),
  modifiedSince = NULL,
  con = NULL
)
}
\arguments{
\item{location}{One ore more USGS station codes.}

\item{parameter_id}{One or more USGS parameter codes. 65 for instantaneous gauge level, 60 for mean daily flow, 61 for instantaneous flow (though beware, these two might be flipped), 10 for water temperature, for example; see more \href{https://help.waterdata.usgs.gov/codes-and-parameters/parameters}{here}.}

\item{start_datetime}{Specify as class Date, POSIXct OR as character string which can be interpreted as POSIXct. If character, UTC offset of 0 will be assigned, otherwise conversion to UTC 0 will be performed on POSIXct class input. If date, time will default to 00:00 to capture whole day.}

\item{end_datetime}{Specify as class Date, POSIXct OR as character string which can be interpreted as POSIXct. If character, UTC offset of 0 will be assigned, otherwise conversion to UTC 0 will be performed on POSIXct class input. If Date, time will default to 23:59:59 to capture whole day.}

\item{modifiedSince}{Optional. A number of hours to narrow the request down to only data points modified within the last x hours. Default NULL fetches all data with the \code{start_datetime} and \code{end_datetime} range.}

\item{con}{A connection to the aquacache database, necessary to allow for the mapping of NWIS approvals, grades, and qualifiers to the database. If left NULL connection will be made and closed automatically.}
}
\value{
A data.frame object of hydrometric data, with datetimes in UTC-0.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

A function used to fetch USGS data from their REST API. See details \href{https://waterservices.usgs.gov/rest/IV-Service.html}{here}. Unit conversions are performed to metric for common parameters like level, flow, temperature.
}
