% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadRWIS.R
\name{downloadRWIS}
\alias{downloadRWIS}
\title{Get data from the RWIS database}
\usage{
downloadRWIS(
  location,
  parameter,
  start_datetime,
  end_datetime = Sys.time(),
  con = NULL,
  rwis = NULL
)
}
\arguments{
\item{location}{Location abbreviation, corresponding to the \code{abbreviation} field of RWDM table 'stations_station'.}

\item{parameter}{One of the rwdm parameter codes, corresponding to a column name in the 'measurements_measurement' table.}

\item{start_datetime}{Specify as class Date, POSIXct OR as character string which can be interpreted as POSIXct. If character, UTC offset of 0 will be assigned, otherwise conversion to UTC 0 will be performed on POSIXct class input. If date, time will default to 00:00 to capture whole day.}

\item{end_datetime}{Specify as class Date, POSIXct OR as character string which can be interpreted as POSIXct. If character, UTC offset of 0 will be assigned, otherwise conversion to UTC 0 will be performed on POSIXct class input. If Date, time will default to 23:59:59 to capture whole day.}

\item{con}{A connection to the aquacache database, necessary to allow for the mapping of NWIS approvals, grades, and qualifiers to the database. If left NULL connection will be made and closed automatically using the function \code{\link[=AquaConnect]{AquaConnect()}} and its default parameters.}

\item{rwis}{A connection to the RWIS database. If left NULL a connection will be made and closed automatically using the function \code{\link[=RWISConnect]{RWISConnect()}} and its default parameters.}
}
\value{
A data.frame object of hydrometric data, with datetimes in UTC-0.
}
\description{
A function used to fetch data from the 'RWDM' database, normally only accessible from within the Yukon Government network.
}
