% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadWSCImages.R
\name{downloadWSCImages}
\alias{downloadWSCImages}
\title{Get images from the WSC}
\usage{
downloadWSCImages(
  location,
  start_datetime,
  username = Sys.getenv("ECCCUSER"),
  password = Sys.getenv("ECCCPASS"),
  url = "https://collaboration.cmc.ec.gc.ca/cmc/hydrometric_additionalData/FieldData/YT/",
  port = NULL,
  save_path = NULL
)
}
\arguments{
\item{location}{The location for which to get images.}

\item{start_datetime}{The earliest datetime to start pulling images from.}

\item{username}{Username to use for password-protected login.}

\item{password}{Password to use for protected login.}

\item{url}{The URL or SFTP server from which to get new images,}

\item{port}{The port to use for SFTP connections.}

\item{save_path}{Optional; path in which to save the image.}
}
\value{
A list object of type 'response' containing the image (object$content) and metadata regarding the object and how/when it was fetched.
}
\description{
Fetches auto images from the WSC/ECCC website or from the NuPoint SFTP website. Default is for ECCC and URL is for the Yukon, adjust for other cameras/locations. To fetch directly from NuPoint specify the SFTP server in the URL, add  the required username/password as variables in the .Renviron file, and point to the with the parameters below. Intended to be called by function \code{\link[=getNewImages]{getNewImages()}}
}
