% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EQConnect.R
\name{EQConnect}
\alias{EQConnect}
\title{Connect to the EQWin database}
\usage{
EQConnect(
  path = "//carver/infosys/EQWin/WR/DB/Water_Resources.mdb",
  silent = FALSE
)
}
\arguments{
\item{path}{Full path to the database including extension.}

\item{silent}{TRUE suppresses messages except for errors.}
}
\value{
A connection to the database.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function exists to facilitate connecting to the EQWin Access database.
}
