% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchFile.R
\name{fetchFile}
\alias{fetchFile}
\title{Extract document stored as BYTEA from postgreSQL DB}
\usage{
fetchFile(
  id,
  id_col,
  ext,
  table,
  con = hydrometConnect(silent = TRUE),
  save_dir = NULL,
  save_name = NULL
)
}
\arguments{
\item{id}{The unique ID from the column specified in \code{id_col} with which to identify the individual record containing the binary object.}

\item{id_col}{The column in which to look for the \code{id}}

\item{ext}{The column in which to look for the file extension, or a file extension (must be preceeded by a period). Only used if save_path is not NULL}

\item{table}{The table to look in for the document.}

\item{con}{A connection to the database.}

\item{save_dir}{A directory in which to write the file.}

\item{save_name}{The name to give the file, \emph{without} extension (specify the extension in \code{ext}).}
}
\value{
A data.frame containing the row(s) identified by the specified ID.
}
\description{
\emph{NOTE} Unless you need flexibility, use functions \code{\link[=fetchDocument]{fetchDocument()}} or \code{\link[=fetchImage]{fetchImage()}}.

Extracts a document/file stored in a postgreSQL database as BYTEA type. Should be flexible enough to work with most schemas, but was designed around the 'documents' and 'images' table created with this package and with files uploaded to the database with \code{\link[=addHydrometDocument]{addHydrometDocument()}} or \code{\link[=addHydrometImageSeries]{addHydrometImageSeries()}}. For extracting rasters to R (which automatically sets the CRS) see \code{\link[rpostgis:pgGetRast]{rpostgis::pgGetRast()}}.
}
