% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchSpatial.R
\name{fetchRaster}
\alias{fetchRaster}
\title{Retrieve raster files from the database}
\usage{
fetchRaster()
}
\value{
An error message telling the user what to do instead.
}
\description{
Placeholder function, not yet complete.
}
