% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fmt}
\alias{fmt}
\title{Format Date-Time to pass to SQL}
\usage{
fmt(x)
}
\arguments{
\item{x}{A date-time object (POSIXct).}
}
\value{
A string formatted as "YYYY-MM-DD HH:MM:SS" in UTC time zone.
}
\description{
Format a date-time object to a string that can be passed to SQL queries. Works on POSIXct objects.
}
