% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/force_update_hydat.R
\name{force_update_hydat}
\alias{force_update_hydat}
\title{Force update of information contained in HYDAT}
\usage{
force_update_hydat(path)
}
\arguments{
\item{path}{The path to the local hydro SQLite database, with extension.}
}
\value{
The database is updated in-place.
}
\description{
By default, the database update functions only update historical information for WSC locations when there is an update to the HYDAT database. However, it may be necessary to refresh this information from time-to-time, for example if any records were accidentally deleted. This function first checks for updates to the HYDAT database on the remote, then replaces locally held information with the HYDAT copy. The function hydro_update_daily is then called to update the daily statistics, as these will be removed in the refresh process. Timeseries held in Aquarius are not affected.
The function also cross-checks any WSC locations present in realtime tables and attempts to find a corresponding historical time-series. Lastly, the table locations will be updated if any new information is found.
}
