% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_free_space_gb}
\alias{get_free_space_gb}
\title{Get free disk space in gigabytes}
\usage{
get_free_space_gb(path = NULL)
}
\arguments{
\item{path}{An optional path to check for free disk space. Defaults to "/" on Unix-like systems and "C:" on Windows.}
}
\value{
The amount of free disk space in gigabytes for the specified path.
}
\description{
Utility function to get the amount of free disk space in gigabytes for a given path. If no path is provided, it defaults to the root directory ("/") on Unix-like systems or the C: drive on Windows. The function uses platform-specific commands to determine free disk space and returns the result in gigabytes.
}
