% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_mappings.R
\name{getImportProfiles}
\alias{getImportProfiles}
\title{List file import profiles}
\usage{
getImportProfiles(con, source_code = NULL, active = TRUE, parse_json = TRUE)
}
\arguments{
\item{con}{A connection to the AquaCache database.}

\item{source_code}{Optional source code filter.}

\item{active}{Optional active filter. Use \code{NULL} to return both active and
inactive profiles.}

\item{parse_json}{If \code{TRUE}, JSON columns are returned as list columns.}
}
\value{
A data.frame of import profiles.
}
\description{
List file import profiles
}
