% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNewHRDPA.R
\name{getNewHRDPA}
\alias{getNewHRDPA}
\title{Get HRDPA rasters}
\usage{
getNewHRDPA(
  start_datetime,
  url = "https://dd.weather.gc.ca/model_hrdpa/2.5km/"
)
}
\arguments{
\item{start_datetime}{The datetime from which to start looking for new rasters}

\item{url}{The url from which to get new rasters.}
}
\value{
A list of lists, where each element consists of the target raster as well as associated attributes.
}
\description{
Get HRDPA rasters
}
