% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNewImages.R
\name{getNewImages}
\alias{getNewImages}
\title{Get new images}
\usage{
getNewImages(image_series_ids = "all", con = NULL, active = "default")
}
\arguments{
\item{image_series_ids}{A vector of image_series_id's. Default 'all' fetches all ids in the table.}

\item{con}{A connection to the database. Leaving NULL will create a connection and close it automatically.}

\item{active}{Sets behavior for import of new images for image series. If set to 'default', the column 'active' in the image_series table will determine whether to get new images or not. If set to 'all', all image series will be fetched regardless of the 'active' column.}
}
\description{
Retrieves images using the active assignment with the lowest fetch priority
in \code{files.image_series_source_adapters} for each image series.
Every source function must have an enabled image-domain entry in
\code{public.source_adapter_capabilities}.
\subsection{Default arguments passed to 'source_fx' functions:}{

This function passes \code{start_datetime}, defaulting to the instant after the
last image. Additional parameters come from the selected assignment's
\code{source_fx_args} JSON object.

If you are a developer, note that download or source functions MUST be registered in AquaCache using function \code{\link[=registerSourceAdapterArguments]{registerSourceAdapterArguments()}}, and that this operation would normally be completed using the 'patch' system. See patch_56.R for examples.
}
}
