% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWSCImages.R
\name{getWSCImages}
\alias{getWSCImages}
\title{Get images from the WSC}
\usage{
getWSCImages(
  location,
  start_datetime,
  username = Sys.getenv("ECCCUSER"),
  password = Sys.getenv("ECCCPASS"),
  url = "https://collaboration.cmc.ec.gc.ca/cmc/hydrometric_additionalData/FieldData/YT/",
  save_path = NULL
)
}
\arguments{
\item{location}{The location for which to get images.}

\item{start_datetime}{The earliest datetime to start pulling images from.}

\item{username}{Username to use for password-protected login.}

\item{password}{Password to use for protected login.}

\item{url}{The URL from which to get new images}

\item{save_path}{Optional; path in which to save the image.}
}
\value{
A list object of type 'response' containing the image (object$content) and metadata regarding the object and how/when it was fetched.
}
\description{
Fetches auto images from the WSC. Default URL is for the Yukon, adjust for other cameras/locations. Intended to be called by function \code{\link[=getNewImages]{getNewImages()}}
}
