% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydat_check.R
\name{hydat_check}
\alias{hydat_check}
\title{Check and update HYDAT}
\usage{
hydat_check(silent = FALSE)
}
\arguments{
\item{silent}{Should messages be printed to the console?}
}
\value{
An updated local copy of hydat, if indicated by the age of the remote HYDAT.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Checks and, if necessary, updates the local version of the WSC HYDAT database. Intended for use on a schedule to ensure the user always has the latest version.
}
