% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_update_daily.R
\name{hydro_update_daily}
\alias{hydro_update_daily}
\title{Daily update of hydro database}
\usage{
hydro_update_daily(path)
}
\arguments{
\item{path}{The path to the hydro database.}
}
\value{
The database is updated in-place.
}
\description{
Daily update of hydro database
}
