% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_update_hourly.R
\name{hydro_update_hourly}
\alias{hydro_update_hourly}
\title{Hourly update of real-time data}
\usage{
hydro_update_hourly(con, tsid = "all")
}
\arguments{
\item{con}{A connection to the database, created with \code{\link[DBI:dbConnect]{DBI::dbConnect()}} or using the utility function \code{\link[WRBtools:hydroConnect]{WRBtools::hydroConnect()}}.}

\item{tsid}{Specific timeseries to update. Default "all" will try to update all timeseries in the database.}
}
\value{
The database is updated in-place, and a data.frame is generated with one row per updated location.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Retrieves new real-time data from WSC and Aquarius, starting from the last data point in the local database. Also updates the location_data_range table. Only works on stations that are ALREADY in the realtime tables; refer to function hydro_update_daily for how to add new stations.

Any timeseries labelled as 'getRealtimeAQ' in the source_fx column in the timeseries table will need your Aquarius username, password, and server address present in your .Renviron profile: see \code{\link[WRBtools:aq_download]{WRBtools::aq_download()}} for more information.
}
