% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_update_weekly.R
\name{hydro_update_weekly}
\alias{hydro_update_weekly}
\title{Weekly update of hydro database}
\usage{
hydro_update_weekly(
  path,
  WSC_range = Sys.Date() - 577,
  aquarius = TRUE,
  aquarius_range = "unapproved",
  server = "https://yukon.aquaticinformatics.net/AQUARIUS"
)
}
\arguments{
\item{path}{The path to the local hydro SQLite database, with extension.}

\item{WSC_range}{The starting date from which to pull real-time WSC data from the web and replace in the local database. Default is max possible days.}

\item{aquarius}{TRUE if you are fetching data from Aquarius, in which case you should also check the next two parameters. FALSE will only replace WSC data.}

\item{aquarius_range}{Should only unapproved (unlocked) data be replaced, or all available data? Select from "all" or "unapproved". Default is "unapproved".}

\item{server}{The URL to your Aquarius server, if needed. Note that your credentials must be in your .Renviron profile: see ?WRBtools::aq_download.}

\item{locations}{The locations you wish to have updated as a character vector. Defaults to "all", though the meaning of all is dependent on the parameter 'aquarius'. Will search for all timeseries with the specified location codes, so level + flow, snow depth + SWE, etc.}
}
\value{
Updated entries in the hydro database.
}
\description{
The weekly update function pulls and replaces a large tranche of Water Survey of Canada and Aquarius data to ensure incorporation of any edits. By default, all real-time water survey data currently available online is replaced, as is any Aquarius data not previously labelled as "approved". Daily means and statistics are recalculated for any potentially affected days in the daily tables, except for daily means provided in HYDAT historical tables.
}
