% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_update_weekly.R
\name{hydro_weekly_update}
\alias{hydro_weekly_update}
\title{Weekly update of hydro database}
\usage{
hydro_weekly_update(
  path,
  WSC_range = Sys.Date() - 577,
  aquarius_range = "unapproved"
)
}
\arguments{
\item{path}{The path to the hydro database.}

\item{WSC_range}{The starting date from which to pull real-time WSC data from the web and replace in the local database. Default is max possible days.}

\item{aquarius_range}{Should only unapproved (locked) data be replaced, or all available data? Select from "all" or "unapproved". Default is "unapproved".}
}
\value{
Updated entries in the hydro database.
}
\description{
The weekly update function pulls and replaces a large tranche of Water Survey of Canada and Aquarius data to ensure incorporation of any edits. By default, all real-time water survey data currently available online is replaced, as is any Aquarius data not previously labelled as "approved"
}
