% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydroConnect.R
\name{hydroConnect}
\alias{hydroConnect}
\title{Connect to hydrology database}
\usage{
hydroConnect(
  name = "hydromet",
  host = "localhost",
  port = "5432",
  username = "postgres",
  password = "SnowFa11ing",
  silent = FALSE
)
}
\arguments{
\item{name}{Database name.}

\item{host}{Database host address.}

\item{port}{Connection port.}

\item{username}{Username. Refrain from using username with write privileges unless you absolutely know what you're doing.}

\item{password}{Password.}

\item{silent}{TRUE suppresses messages except for errors.}
}
\value{
A connection to the database.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function exists to facilitate connecting to the hydrology database, especially if the database type and connection method changes in the future. Cannot be used to create a new database, use \code{\link[DBI:dbConnect]{DBI::dbConnect()}} with the appropriate driver (e.g. \code{\link[RSQLite:SQLite]{RSQLite::SQLite()}}) for that purpose.
}
\seealso{
\code{\link[=DB_browse_ts]{DB_browse_ts()}}, \code{\link[=DB_browse_spatial]{DB_browse_spatial()}}to browse the database contents; \code{\link[=DB_get_meta]{DB_get_meta()}} to extract metadata; \code{\link[=DB_get_ts]{DB_get_ts()}} to extract timeseries information; and \code{\link[=DB_get_spatial]{DB_get_spatial()}} to extract spatial content.
}
