% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrometApp.R
\name{hydrometApp}
\alias{hydrometApp}
\title{Shiny Interface to Hydromet DB}
\usage{
hydrometApp()
}
\value{
Opens the Shiny application.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Loads a Shiny app that facilitates the addition of timeseries (data, images, and rasters) and uploading of documents, one-off images and rasters.
}
