% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeMissing.R
\name{imputeMissing}
\alias{imputeMissing}
\title{Impute missing values}
\usage{
imputeMissing(
  tsid,
  radius,
  start,
  end,
  extra_params = NULL,
  imputed = TRUE,
  con = hydrometConnect(silent = TRUE)
)
}
\arguments{
\item{tsid}{The target timeseries_id.}

\item{radius}{The radius in kilometers within which to search for similar timeseries.}

\item{start}{The start datetime (as POSIXct object) from which to fill in missing values (can use local or other time zone).}

\item{end}{The end datetime (as POSIXct object) to which to fill in missing values (can use local or other time zone).}

\item{extra_params}{Optional extra parameters to consider when imputing values. For example, you may choose to use wind speeds at 1 or 10 meters to impute speeds at 5 meters.}

\item{imputed}{Should already imputed data be imputed again?}

\item{con}{A connection to the database.}
}
\value{
Imputed values added to the database.
}
\description{
Impute missing values using either nearby timeseries or using linear or spline interpollation. If using nearby timeseries, they can be of same or of other parameters, as specified with parameter 'extra_params'. User interaction is necessary to select the most appropriate timeseries and to review the result of imputation before updating the database.
}
