% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{inf_to_na}
\alias{inf_to_na}
\title{Replace infinite or NaN values with NA}
\usage{
inf_to_na(x)
}
\arguments{
\item{x}{Numeric vector, data.frame, or data.table. If data.frame or data.table, will only work on 'numeric' class columns.}
}
\value{
Numeric vector with infinite values converted to \code{NA}.
}
\description{
Utility function to replace \code{Inf}, \code{-Inf}, and \code{NaN} values with \code{NA}.
}
