% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_create_tables.R
\name{initial_create}
\alias{initial_create}
\title{Initial hydro database creation.}
\usage{
initial_create(path)
}
\arguments{
\item{path}{The path to the existing database, or the path where the SQlite database should be created.}
}
\value{
An SQLite database in the folder location specified by 'path'.
}
\description{
Creates an SQLite database or replaces an existing database. Established pre-set table structure and populates defaults in the "settings" table.
}
