% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_initial_WSC.R
\name{initial_WSC}
\alias{initial_WSC}
\title{Initial creation of tables containing WSC data}
\usage{
initial_WSC(path)
}
\arguments{
\item{path}{The path to the local hydro database.}
}
\value{
Updated tables in the database.
}
\description{
Pulls data from Aquarius as well as from WSC to perform the initial creation of WSC-related tables. The incorporation of Aquarius data is done to fill the gap between the 18 months of real-time data and the published historical data (daily means), as well as to provide finer resolution data.
}
\details{
Uses two functions under the hood that require some set-up of the .Renviron file: tidyhydat.ws::realtime_ws and WRBtools::aq_download. See respective help files for setup information.
}
