% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertACImage.R
\name{insertACImage}
\alias{insertACImage}
\title{Add images to aquacache database}
\usage{
insertACImage(
  object,
  datetime,
  image_type,
  fetch_datetime = NULL,
  img_series_id = NULL,
  description = NULL,
  tags = NULL,
  owner = NULL,
  contributor = NULL,
  share_with = "public_reader",
  location = NULL,
  latitude = NULL,
  longitude = NULL,
  azimuth_true = NULL,
  elevation_agl = NULL,
  elevation_msl = NULL,
  con = NULL
)
}
\arguments{
\item{object}{Valid path including extension to the image to upload, or an object of class 'response' such as that provided by \code{\link[=downloadWSCImages]{downloadWSCImages()}}.}

\item{datetime}{The datetime the image was taken at, as a POSIXct object or something that can be coerced to one. If not POSIXct timezone is assumed to be UTC.}

\item{image_type}{The image_type_id from table images_types corresponding to the image type. Pass as numeric, must match an entry in table 'image_types'}

\item{fetch_datetime}{The datetime the image was retrieved (optional). If not POSIXct timezone is assumed to be UTC.}

\item{img_series_id}{The img_series_id, from the table image_series, corresponding to the image location and type. Set NULL if the image is not linked to an image series - in that case you'll need to specify, at minimum, a latitude and longitude for the image OR associate it with a location.}

\item{description}{A description of the image. Pass as text, can be left NULL; consider also using the \code{tags} parameter.}

\item{tags}{Tags to associate with the image. Pass as a character vector, one element per tag.}

\item{owner}{The owner of the image, matching the organization_id of table 'organizations'. Can be left NULL.}

\item{contributor}{The contributor of the image, matching the organization_id of table 'organizations'. Pass as text, can be left NULL.}

\item{share_with}{Which user groups should the image be shared with. Default is 'public_reader', the public group. Pass as a character vector, one element per group.}

\item{location}{The location_id with which to associate the document (must be in the database). Pass a location_id as a numeric. If img_series_id is specified, the location already associated is used. This parameter can also be left NULL if latitude and longitude are specified.}

\item{latitude}{If no img_series_id exists yet AND not specifying a location: the latitude of the image. Pass as numeric.}

\item{longitude}{If no img_series_id exists yet AND not specifying a location: the longitude of the image. Pass as numeric.}

\item{azimuth_true}{Optional: the true azimuth of the image (direction the image was taken in. Pass as numeric.}

\item{elevation_agl}{Optional: the elevation above ground level of the image in METERS. Pass as numeric.}

\item{elevation_msl}{Optional: the elevation above mean sea level of the image in METERS. Pass as numeric.}

\item{con}{A connection to the database. Default NULL uses AquaConnect() and close the connection afterwards.}
}
\value{
TRUE if an image was properly added to the database.
}
\description{
Images auto-generated on a regular basis should be updated in the database using function \code{\link[=getNewImages]{getNewImages()}} instead of this function. For one-off images, read on.

This function facilitates the addition of one image at a time to the database in the 'images' table. Each image must be linked to a specific location_id from the locations table. Adding an image directly to the database is not possible since the file must be converted to a binary object before loading. See \code{\link[YGwater:getImage]{YGwater::getImage()}} to get the image out again.
}
