% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertACVector.R
\name{insertACVector}
\alias{insertACVector}
\title{Add a vector file to the database}
\usage{
insertACVector(
  geom,
  layer_name,
  feature_name = NULL,
  description = NULL,
  feature_name_col = NULL,
  description_col = NULL,
  table = "vectors",
  schema = "spatial",
  geom_col = "geom",
  overwrite = FALSE,
  ask = TRUE,
  con = NULL
)
}
\arguments{
\item{geom}{The geometry object to add to the database, as a \code{\link[terra:vect]{terra::vect()}} object or as a file path to a shapefile, geopackage or something else that terra::vect() van use. Conversion will automatically be made to epsg:4269, NAD83 lat/long decimal degrees. Can be points, lines, or polygons with one or more features. Multi-feature geoms will be split up into individual database entries.}

\item{layer_name}{The name to give to the vector layer, which defines which layer_name it gets assigned to in the database. This should always be an existing layer_name unless you have a good reason to create a new one.}

\item{feature_name}{A short but descriptive name to give to the geom feature. Leave NULL if specified with parameter \code{feature_name_col} instead. This parameter only works for geoms with a single feature (row).}

\item{description}{Optional but highly recommended long-form description of the geometry feature. Leave NULL is specifying a \code{description_col} instead.}

\item{feature_name_col}{The name of the column with names to give to the geom features. Each feature (row in the attribute table) will be entered to the database using the string in this column. Leave NULL if specified with parameter \code{feature_name}.}

\item{description_col}{The name of the column containing descriptions associated with each feature. Each feature (row in the attribute table) will be entered to the database using the string in this column. Leave NULL if specified with parameter \code{description} instead.}

\item{table}{The target table in the database (as character string). See 'schema' if not under the 'spatial' schema.}

\item{schema}{The schema in which the target 'table' is located. Default is 'spatial'. Note that this is NOT the default for \code{\link[rpostgis:pgWriteGeom]{rpostgis::pgWriteGeom()}}.}

\item{geom_col}{The name of the database table column in which to insert the geometry object.}

\item{overwrite}{If a row already exists for the combination of layer_name, name,  and geometry type (point, line, or polygon), should it be overwritten?}

\item{ask}{Whether to ask for user confirmation when creating a new layer_name or overwriting existing entries. Default TRUE.}

\item{con}{A connection to the database. Default NULL will use the utility function \code{\link[=AquaConnect]{AquaConnect()}} and disconnect afterwards.}
}
\value{
A boolean vector, one element per feature.Messages will also be printed to the console.
}
\description{
Use this function to add a vector file to the database. Ensures that database constraints are met and permits spatial queries. If you need to delete (not overwrite) a vector for any reason you'll have to use SQL (perhaps via R using the DBI package).
\subsection{Extracting from the database}{

Use function \code{\link[YGwater:getVector]{YGwater::getVector()}} to retrieve a point, line, or polygon from the database.
}

\subsection{Attribute tables:}{

The attribute table of the object will be retained in the \code{attributes} JSON column for all fields that are not already
represented by dedicated columns in the \code{vectors} table. This preserves additional metadata for each feature while still
allowing spatial relationships to be maintained within the database. Fields supplied through \code{feature_name}/\code{feature_name_col}
and \code{description}/\code{description_col} are still written to their respective columns, with all remaining fields stored as JSON.
}
}
