% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertHydrometImage.R
\name{insertHydrometImage}
\alias{insertHydrometImage}
\title{Add images to hydrometric database}
\usage{
insertHydrometImage(
  object,
  img_meta_id,
  datetime,
  fetch_datetime = NULL,
  location = NULL,
  image_type = NULL,
  con = hydrometConnect()
)
}
\arguments{
\item{object}{Valid path including extension to the document to upload, or an object of class 'response' such as that provided by \code{\link[=getNewWSCImages]{getNewWSCImages()}}.}

\item{img_meta_id}{The img_meta_id, from the table images_index, corresponding to the image location and type. Set NULL if there is no img_meta_id yet.}

\item{datetime}{The datetime the image was taken at, as a POSIXct object or something that can be coerced to one. If not POSIXct timezone is assumed to be UTC.}

\item{fetch_datetime}{The datetime the image was retrieved (optional).}

\item{location}{If no img_meta_id exists yet: the location with which to associate the document (must be in the database).}

\item{image_type}{If no img_meta_id exists yet: the type of image: 'auto', or 'manual'.}

\item{con}{A connection to the database, created with \code{\link[DBI:dbConnect]{DBI::dbConnect()}} or using the utility function \code{\link[=hydrometConnect]{hydrometConnect()}}.}
}
\value{
TRUE if an image was properly added to the database.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Images auto-generated on a regular basis should be updated in the database using function \code{\link[=getNewImages]{getNewImages()}} instead of this function. For one-off images, read on.

This function facilitates the addition of one image at a time to the database in the 'images' table. Each image must be linked to a specific location. Adding an image directly to the database is not possible, since the file must be converted to a binary object before loading. See \code{\link[=fetchImage]{fetchImage()}} to get the image out again.
}
