% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertVector.R
\name{insertVector}
\alias{insertVector}
\title{Add a vector file to the database}
\usage{
insertVector(con, geom, table, geom_col_name = "geom")
}
\arguments{
\item{con}{A connection to the database, created with \code{\link[DBI:dbConnect]{DBI::dbConnect()}} or using the utility function \code{\link[=hydrometConnect]{hydrometConnect()}}.}

\item{geom}{The geometry object to add to the database, as a \code{\link[terra:vect]{terra::vect()}} object. Conversion will automatically be made to epsg:4269, NAD83 lat/long decimal degrees. Can be points, lines, or polygons.}

\item{table}{The referenced table in the database (as character string), must be under the public schema.}

\item{geom_col_name}{The name of the column in which to insert the geometry object.}
}
\value{
TRUE if the vector was added.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Use this function to add a vector file such as a drainage polygon, provincial buffer, waterbodies, etc to the database. Ensures that database constraints are met. If you need to replace or delete a vector for any reason you'll have to use SQL (perhaps via R using the DBI package) to delete it first or using \code{\link[rpostgis:pgWriteGeom]{rpostgis::pgWriteGeom()}} with the necessary parameters.
}
\section{If adding polygons:}{
Polygons must have an attribute table with columns for 'polygon_type', 'description', 'name'. polygon_type column can only have values in c('drainage_basin', 'buffer', 'waterbody', 'prov_terr', 'country', 'other')
}

\section{If adding points:}{
Adding points is not supported yet.
}

\section{if adding lines:}{
Adding lines is not supported yet.
}

