% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSnowTemplate.R
\name{readSnowTemplate}
\alias{readSnowTemplate}
\title{Read snow template and import into Snow database}
\usage{
readSnowTemplate(template)
}
\arguments{
\item{template}{The name of the template (excel workbook) containing the snow data.}
}
\value{
Does not return any object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function exists to facilitate adding new snow survey data to the snow database. It reads the snow template workbook and imports it directly into the snow database.
}
