% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RWISConnect.R
\name{RWISConnect}
\alias{RWISConnect}
\title{Connect to the RWIS database}
\usage{
RWISConnect(
  name = Sys.getenv("rwisName", "rwdm"),
  host = Sys.getenv("rwisHost", "rwis.gov.yk.ca"),
  port = Sys.getenv("rwisPort", "5432"),
  username = Sys.getenv("rwisUser"),
  password = Sys.getenv("rwisPass")
)
}
\arguments{
\item{name}{Database name.}

\item{host}{Database host address.}

\item{port}{Connection port.}

\item{username}{Username. Refrain from using username with write privileges unless you absolutely know what you're doing.}

\item{password}{Password.}
}
\value{
A connection to the database
}
\description{
Establishes a connection to the RWIS (Road Weather Information System) PostgreSQL database. Only works from within YG networks.
}
\seealso{
\code{\link[=AquaConnect]{AquaConnect()}} for establishing a connection to the WRB's hydrometric database.
}
