% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowConnect.R
\name{snowConnect}
\alias{snowConnect}
\title{Connect to the snow postgres database}
\usage{
snowConnect(
  name = "snowDB",
  host = Sys.getenv("snowHost"),
  port = Sys.getenv("snowPort"),
  username = Sys.getenv("snowAdminUser"),
  password = Sys.getenv("snowAdminPass"),
  silent = FALSE
)
}
\arguments{
\item{name}{Database name.}

\item{host}{Database host address.}

\item{port}{Connection port.}

\item{username}{Username. Refrain from using username with write privileges unless you absolutely know what you're doing.}

\item{password}{Password.}

\item{silent}{TRUE suppresses messages except for errors.}
}
\value{
A connection to the database.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function exists to facilitate connecting to the snow database, especially if the database type and connection method changes in the future. This function is different from the \code{\link[YGwater:snowConnect]{YGwater::snowConnect()}} function in that it defaults to an account with write privileges.
}
