% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowConnect_access.R
\name{snowConnect_access}
\alias{snowConnect_access}
\title{Connect to snow survey database}
\usage{
snowConnect_access(path = "default", silent = FALSE)
}
\arguments{
\item{path}{The path to the database. Currently supports either a Microsoft Access database or an .sqlite database. 'default' points to //carver/infosys/Snow/DB/SnowDB.mdb.}

\item{silent}{TRUE suppresses messages except for errors.}
}
\value{
A connection to the database.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Connects to the old Snow Survey Access database.To connect to the new postgres database, use \code{\link[=snowConnect]{snowConnect()}}.
}
