% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowConnect.R
\name{snowConnect_pg}
\alias{snowConnect_pg}
\title{Connect to the snow database}
\usage{
snowConnect_pg(
  name = "snowDB",
  host = "localhost",
  port = "5432",
  username = "postgres",
  password = Sys.getenv("SnowPass"),
  silent = FALSE
)
}
\arguments{
\item{name}{Database name.}

\item{host}{Database host address.}

\item{port}{Connection port.}

\item{username}{Username. Refrain from using username with write privileges unless you absolutely know what you're doing.}

\item{password}{Password.}

\item{silent}{TRUE suppresses messages except for errors.}
}
\value{
A connection to the database.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function exists to facilitate connecting to the snow database, especially if the database type and connection method changes in the future.
}
