% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowInit.R
\name{snowInit}
\alias{snowInit}
\title{Initial PostgreSQL snow database creation.}
\usage{
snowInit(con = snowConnect_pg(), overwrite = FALSE)
}
\arguments{
\item{con}{A connection to the database, created with \code{\link[DBI:dbConnect]{DBI::dbConnect()}} or using the utility function \code{\link[=snowConnect_pg]{snowConnect_pg()}}.}

\item{overwrite}{TRUE overwrites the database, if one exists in the same path. Nothing will be kept. FALSE will create tables only where they are missing.}
}
\value{
A PostgreSQL database in ....
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Creates a PostgreSQL database or replaces an existing database. Established pre-set table structure. All tables are created and with primary keys.
}
