% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synchronize.R
\name{synchronize}
\alias{synchronize}
\title{Synchronize hydro DB with remote sources}
\usage{
synchronize(
  con = hydrometConnect(silent = TRUE),
  timeseries_id = "all",
  start_datetime
)
}
\arguments{
\item{con}{A connection to the database, created with \code{\link[DBI:dbConnect]{DBI::dbConnect()}} or using the utility function \code{\link[=hydrometConnect]{hydrometConnect()}}.}

\item{timeseries_id}{The timeseries_ids you wish to have updated, as character or numeric vector. Defaults to "all".}

\item{start_datetime}{The datetime (as a POSIXct) from which to look for possible new data. You can specify a single start_datetime to apply to all \code{timeseries_id}, or one per element of \code{timeseries_id.}}
}
\value{
Updated entries in the hydro database.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

The synchronize function pulls and replaces data of category 'continuous' and 'discrete' if and when a discrepancy is observed between the remote repository and the local data store, with the remote taking precedence. For continuous data daily means and statistics are recalculated for any potentially affected days in the daily tables, except for daily means provided in HYDAT historical tables.

NOTE that any data point labelled as imputed = TRUE is only replaced if a value is found in the remote.

Any timeseries labelled as 'getRealtimeAquarius' in the source_fx column in the timeseries table will need your Aquarius username, password, and server address present in your .Renviron profile: see \code{\link[=getRealtimeAquarius]{getRealtimeAquarius()}} for more information.
}
