% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_hydat_datums.R
\name{update_hydat_datums}
\alias{update_hydat_datums}
\title{Update datum information with HYDAT database entries}
\usage{
update_hydat_datums(con = hydrometConnect())
}
\arguments{
\item{con}{A connection to the database, created with \code{\link[DBI:dbConnect]{DBI::dbConnect()}} or using the utility function \code{\link[=hydrometConnect]{hydrometConnect()}}.}
}
\value{
HYDAT-related information is updates in place in the DB.
}
\description{
Update datum information with HYDAT database entries
}
