% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateACDocument.R
\name{updateHydrometDocument}
\alias{updateHydrometDocument}
\title{Update existing document in aquacache database}
\usage{
updateHydrometDocument(
  id,
  new_path = NULL,
  new_name = NULL,
  new_type = NULL,
  new_description = NULL,
  new_authors = NULL,
  new_publish_date = NULL,
  new_url = NULL,
  new_user_groups = NULL,
  new_geoms = NULL,
  con = AquaConnect(),
  check = TRUE
)
}
\arguments{
\item{id}{The document_id of the document to update, from the document_id column of the documents table.}

\item{new_path}{Valid path including extension to the document to upload.}

\item{new_name}{A concise but descriptive name to give the document.}

\item{new_type}{Type of document, which must exist in the database already. Currently one of 'thesis', 'report', 'well log', 'conference paper', 'poster', 'journal article', 'map', 'graph', 'protocol', 'metadata', 'audit'.}

\item{new_description}{A text description of what the document is. Please be detailed!}

\item{new_authors}{Document author(s) if known. Specify multiple authors as individual elements of a character vector, such as c("author 1", "author 2").}

\item{new_publish_date}{The date of publication, as a Date object.}

\item{new_url}{An optional url (could also be a DOI) for the document.}

\item{new_user_groups}{New user groups to associate with the document. Leave NULL for no change.}

\item{new_geoms}{The geom_id(s) with which to associate the document (must be in the database table 'vectors'). Leave NULL for a document with no spatial context. Will overwrite any existing associations.}

\item{con}{A connection to the database, created with \code{\link[DBI:dbConnect]{DBI::dbConnect()}} or using the utility function \code{\link[=AquaConnect]{AquaConnect()}}.}

\item{check}{Logical, whether to double check that we're working with the right document before proceeding. Default is TRUE}
}
\value{
TRUE if a document was properly added to the database.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function facilitates the updating of one document at a time to the database in the 'documents' table. Adding a document directly to the database is not possible, since the file must be converted to a binary object before loading. See \code{\link[YGwater:getDocument]{YGwater::getDocument()}} to get a document out again.
\subsection{Locations, lines, and polygons}{

Any document can be associated with locations (points), lines, polygons, or any combination thereof. Please reference the table 'vectors' to give the correct geom_id(s) for your desired geoms.
}
}
