% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateACUser.R
\name{validateACUser}
\alias{validateACUser}
\title{Validate a user's credentials against those in the database.}
\usage{
validateACUser(username, password, con = AquaConnect())
}
\arguments{
\item{username}{The username for the user to validate.}

\item{password}{The password for the user to validate.}

\item{con}{A connection to the AquaCache database. Default uses \code{\link[=AquaConnect]{AquaConnect()}}.}
}
\value{
TRUE if the user's credentials are valid, FALSE otherwise.
}
\description{
Validate a user's credentials against those in the database.
}
