% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WRBdatabase-package.R
\docType{package}
\name{WRBdatabase-package}
\alias{WRBdatabase}
\alias{WRBdatabase-package}
\title{WRBdatabase: Hydrological database creation and maintenance}
\description{
Set of functions for the initial creation and ongoing maintenance of WRB's local database. This local database is a combination of the Water Survey of Canada real-time and historical data and WRB data held in Aquarius. Supports flow and level parameters from WSC and any consistently-named timeseries from Aquarius. Multiple parameters are calculated to facilitate reporting and graphically representing data.
}
\author{
\strong{Maintainer}: Ghislain de Laplante \email{ghislaindel@gmail.com} (\href{https://orcid.org/0000-0002-5093-9185}{ORCID})

}
\keyword{internal}
