% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addACImageSeries.R
\name{addACImageSeries}
\alias{addACImageSeries}
\title{Add new image series (auto-fetched)}
\usage{
addACImageSeries(
  location,
  start_datetime,
  source_fx,
  source_fx_args = NA,
  share_with = "public_reader",
  visibility_public = "exact",
  con = NULL
)
}
\arguments{
\item{location}{The location or location_id associated with the image series. Pass a location code as text and a location_id as a numeric.}

\item{start_datetime}{The datetime (as POSIXct) from which to look for images}

\item{source_fx}{The function to use for fetching new images. Must be an existing function in this package.}

\item{source_fx_args}{Arguments to pass to the function(s) specified in parameter 'source_fx'. See details.}

\item{share_with}{A \emph{character} vector of the user group(s) with which to share the timeseries, Default is 'public_reader'. Pass multiple groups as a single string, e.g. "public_reader, YG"}

\item{visibility_public}{How should the image location be publicly visible? Options are 'exact', 'region', 'jitter'.}

\item{con}{A connection to the database, created with \code{\link[DBI:dbConnect]{DBI::dbConnect()}} or using the utility function \code{\link[=AquaConnect]{AquaConnect()}}. If left NULL, a connection will be attempted using AquaConnect() and closed afterwards.}
}
\value{
TRUE if successful, and a new entry in the database with images fetched.
}
\description{
Use this function to add a new image series, i.e. a continually updating time-series of images that can be accessed from the web or internal server location on a regular basis.
}
\details{
#' Additional arguments to pass to the function specified in \code{source_fx} go in argument \code{source_fx_args} and will be converted to JSON format. It's therefore necessary to pass this argument in as a single length character vector in the style "argument1: value1, argument2: value2".
}
