% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addACImageSeries.R
\name{addACImageSeries}
\alias{addACImageSeries}
\title{Add new image series (auto-fetched)}
\usage{
addACImageSeries(
  location,
  start_datetime,
  source_fx,
  source_fx_args = NA,
  share_with = 1,
  visibility_public = "exact",
  con = AquaConnect(silent = TRUE)
)
}
\arguments{
\item{location}{The location or location_id associated with the image series. Pass a location code as text and a location_id as a numeric.}

\item{start_datetime}{The datetime (as POSIXct) from which to look for images}

\item{source_fx}{The function to use for fetching new images. Must be an existing function in this package.}

\item{source_fx_args}{Additional arguments to pass to the function, in the form "\{param1 = arg1\}, \{param2 = 'arg2'\}". Each parameter = value pair needs to be enclosed in curly brackets, which might be missing here. Do not deviate from this format!}

\item{share_with}{Which user groups should the image series be shared with? Default is '1', the public group.}

\item{visibility_public}{How should the image location be publicly visible? Options are 'exact', 'region', 'jitter'.}

\item{con}{A connection to the database, created with \code{\link[DBI:dbConnect]{DBI::dbConnect()}} or using the utility function \code{\link[=AquaConnect]{AquaConnect()}}.}
}
\value{
TRUE if successful, and a new entry in the database with images fetched.
}
\description{
Use this function to add a new image series, i.e. a continually updating time-series of images that can be accessed from the web or internal server location on a regular basis.
}
