% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addACLocation.R
\name{addACLocation}
\alias{addACLocation}
\title{Add location to aquacache}
\usage{
addACLocation(
  df = NULL,
  name = NA,
  name_fr = NA,
  alias = NA,
  location_code = NA,
  latitude = NA,
  longitude = NA,
  share_with = NA,
  location_type = NA,
  note = NA,
  contact = NA,
  datum_id_from = NA,
  datum_id_to = NA,
  conversion_m = NA,
  current = NA,
  network = NA,
  project = NA,
  con = NULL
)
}
\arguments{
\item{df}{A data.frame containing the location fields. \code{datum_id_from},
\code{datum_id_to}, \code{conversion_m}, and \code{current} are optional. When
\code{conversion_m} is missing, \code{\link[=get_elevation]{get_elevation()}} estimates the elevation from
\code{latitude} and \code{longitude} and the matching datum IDs are populated
automatically. If this parameter is provided, all other parameters except
for \code{con} must be left as their default values.}

\item{name}{A character vector of the location name(s).}

\item{name_fr}{A character vector of the location name(s) in French. You're highly encouraged to populate this field, but if left blank (or the corresponding column in \code{df} is missing or empty) it will be populated with 'Traduction requise!'.}

\item{alias}{A character vector of the location alias(es). This is optional, leave NA if not needed.}

\item{location_code}{A character vector of the location code(s). Note that in most cases this should be left NULL to auto-generate, or created when adding a new location using the YGwater Shiny application!}

\item{latitude}{A numeric vector of the latitude(s) as decimal degrees.}

\item{longitude}{A numeric vector of the longitude(s) as decimal degrees.}

\item{share_with}{A character vector of the user group(s) with which to share the location(s), separated by a comma. Default public group is "public_reader".}

\item{location_type}{A numeric vector of the location type(s) id(s) from table 'location_types'.}

\item{note}{A character vector of notes for the location(s) (optional).}

\item{contact}{A character vector of the contact(s) for the location(s) (optional).}

\item{datum_id_from}{A numeric vector of the datum ID(s) from which the
location(s) are measured, from \code{public.datum_list}. When \code{conversion_m} is
omitted, this is set to the Assumed Datum ID.}

\item{datum_id_to}{A numeric vector of the datum ID(s) to which the
location(s) are measured, from \code{public.datum_list}. When \code{conversion_m} is
omitted, this is matched to the vertical datum returned by
\code{\link[=get_elevation]{get_elevation()}}. A missing datum is added to \code{public.datum_list}.}

\item{conversion_m}{A numeric vector of conversion factors from
\code{datum_id_from} to \code{datum_id_to}. Missing values are estimated with
\code{\link[=get_elevation]{get_elevation()}}.}

\item{current}{A logical vector indicating whether the conversion factor(s)
are current. Missing values default to \code{TRUE}.}

\item{network}{A numeric vector of the network(s) to which the location(s) belong.}

\item{project}{A numeric vector of the project(s) to which the location(s) belong.}

\item{con}{A connection to the aquacache database. Default uses \code{\link[=AquaConnect]{AquaConnect()}}. If left NULL the function will attempt to connect to the database and automatically disconnect afterwards.}
}
\value{
Invisibly, a data.frame describing the new locations and their
datum conversions.
}
\description{
Adds a new location to the aquacache 'locations' table. You can pass a data.frame with the necessary columns, or provide each parameter separately. Extensive checks are performed to ensure that the location does not already exist, and that all necessary parameters are provided and are valid.
}
