% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addACLocation.R
\name{addACLocation}
\alias{addACLocation}
\title{Add location to AquaCache}
\usage{
addACLocation(
  df = NULL,
  location = NA,
  name = NA,
  name_fr = NA,
  latitude = NA,
  longitude = NA,
  visibility_public = NA,
  share_with = 1,
  owner = NA,
  data_sharing_agreement_id = NA,
  location_type = NA,
  note = NA,
  contact = NA,
  datum_id_from = NA,
  datum_id_to = NA,
  conversion_m = NA,
  current = NA,
  network = NA,
  project = NA,
  con = NULL
)
}
\arguments{
\item{df}{A data.frame containing the following columns: location, name, name_fr, latitude, longitude, visibility_public, share_with, owner, data_sharing_agreement_id, location_type, note, contact, datum_id_from, datum_id_to, conversion_m, current, network, project. If this parameter is provided, all other parameters must be NA or left as their default values.}

\item{location}{A character vector of the location code(s).}

\item{name}{A character vector of the location name(s).}

\item{name_fr}{A character vector of the location name(s) in French.}

\item{latitude}{A numeric vector of the latitude(s).}

\item{longitude}{A numeric vector of the longitude(s).}

\item{visibility_public}{A character vector of the visibility of the location(s). Default is 'exact'.}

\item{share_with}{A numeric vector of the user group(s) with which to share the location(s). Default is 1.}

\item{owner}{A numeric vector of the owner(s) of the location(s).}

\item{data_sharing_agreement_id}{A numeric vector of the data sharing agreement(s) for the location(s).}

\item{location_type}{A numeric vector of the location type(s).}

\item{note}{A character vector of notes for the location(s).}

\item{contact}{A character vector of the contact(s) for the location(s).}

\item{datum_id_from}{A numeric vector of the datum ID(s) from which the location(s) are measured.}

\item{datum_id_to}{A numeric vector of the datum ID(s) to which the location(s) are measured.}

\item{conversion_m}{A numeric vector of the conversion factor(s) from the datum_id_from to the datum_id_to.}

\item{current}{A logical vector of whether the conversion factor(s) are current.}

\item{network}{A numeric vector of the network(s) to which the location(s) belong.}

\item{project}{A numeric vector of the project(s) to which the location(s) belong.}

\item{con}{A connection to the AquaCache database. Default uses \code{\link[=AquaConnect]{AquaConnect()}}. If left NULL the function will attempt to connect to the database and automatically disconnect afterwards.}
}
\value{
Success/error messages and new entries added to the database.
}
\description{
Adds a new location to the AquaCache 'locations' table. You can pass a data.frame with the necessary columns, or provide each parameter separately. Extensive checks are performed to ensure that the location does not already exist, and that all necessary parameters are provided and are valid.
}
