% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addACRasterSeries.R
\name{addACRasterSeries}
\alias{addACRasterSeries}
\title{Add new raster series (auto-fetched models)}
\usage{
addACRasterSeries(
  model,
  parameter,
  start_datetime,
  source_fx,
  type,
  source_fx_args = NA,
  con = NULL,
  media_id = NA,
  matrix_state_id = NA,
  aggregation_type = NA,
  z_value = NA,
  z_units = NA
)
}
\arguments{
\item{model}{The model producing the raster, as a character string}

\item{parameter}{The parameter to import, as a parameter name or \code{parameter_id}.}

\item{start_datetime}{The datetime (as POSIXct) from which to look for rasters}

\item{source_fx}{The function to use for fetching new rasters. Must be an existing function in this package.}

\item{type}{The type of raster, 'forecast' or 'reanalysis'. Reanalysis rasters are kept forever, forecasts are replaced when a new one is issued.}

\item{source_fx_args}{Arguments to pass to the function(s) specified in parameter 'source_fx'. See details.}

\item{con}{A connection to the database, created with \code{\link[DBI:dbConnect]{DBI::dbConnect()}}. Leave NULL if you want to use the package default connection settings and have the connection automatically closed afterwards.}

\item{media_id}{Optional \code{media_id} for the raster series. Required when the parameter can exist in multiple matrix states unless \code{matrix_state_id} is supplied explicitly.}

\item{matrix_state_id}{Optional \code{matrix_state_id} for the raster series. Leave as \code{NA} to let the database resolve it from \code{media_id} and the parameter defaults.}

\item{aggregation_type}{Optional aggregation type name or \code{aggregation_type_id} for the raster series metadata.}

\item{z_value}{Optional z value (elevation, depth, pressure level, etc.) associated with the raster series.}

\item{z_units}{Optional units for \code{z_value}.}
}
\value{
TRUE if successful, and a new entry in the database with images fetched.
}
\description{
Use this function to add a new series of rasters generated by a model, such as the RDPS or HRDPS models by ECCC.
}
\details{
Additional arguments to pass to the function specified in source_fx go in argument 'source_fx_args' (or a column with same name in 'df') and will be converted to JSON format. It's therefore necessary to pass this argument in as a single length character vector in the style "argument1: value1, argument2: value2".
}
