% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addACUserGroup.R
\name{addACUserGroup}
\alias{addACUserGroup}
\title{Add a new user group to the aquacache}
\usage{
addACUserGroup(name, description, con = NULL)
}
\arguments{
\item{name}{The name of the new user group}

\item{description}{A description of the new user group}

\item{con}{A connection to the database, created with \code{\link[DBI:dbConnect]{DBI::dbConnect()}} or using the utility function \code{\link[=AquaConnect]{AquaConnect()}}. NULL will create a connection and close it afterwards, otherwise it's up to you to close it after.}
}
\value{
The group_id of the new user group, plus a new entry to the database.
}
\description{
Add a new user group to the aquacache
}
