% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addNewContinuous.R
\name{addNewContinuous}
\alias{addNewContinuous}
\title{Add new data to the measurements_continuous table}
\usage{
addNewContinuous(tsid, df, con = NULL, overwrite = "no")
}
\arguments{
\item{tsid}{The timeseries_id to which the data will be appended. This is a required parameter.}

\item{df}{A data.frame containing the data. Must have columns named 'datetime' and 'value' at minimum. Other optional columns are 'owner', 'contributor', 'approval', 'grade', 'qualifier', 'data_sharing_agreement_id', 'imputed'; see the \code{adjust_} series of functions to see how these are used.}

\item{con}{A connection to the database, created with \code{\link[DBI:dbConnect]{DBI::dbConnect()}} or using the utility function \code{\link[=AquaConnect]{AquaConnect()}}. If left NULL, a connection will be attempted using AquaConnect() and closed afterwards.}

\item{overwrite}{Select one of "no", "all", "conflict". Default is "no", which will not overwrite any existing data (and fail if there is a conflict). "all" will wipe and replace all database entries for the target timeseries in the temporal range of \code{df}, while "conflict" will overwrite only those entries that conflict with the data in \code{df} (i.e. have the same datetime or date).}
}
\value{
Nothing; data is added to the database silently.
}
\description{
This function can be used to append new continuous type data directly to the database without downloading it from a remote source. Differs from \code{\link[=getNewContinuous]{getNewContinuous()}} as the latter is used to pull data from a remote before append. New data can only be appended to basic timeseries.
}
