% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addNewContinuous.R
\name{addNewContinuous}
\alias{addNewContinuous}
\title{Add new data to the measurements_continuous or measurements_calculated_daily tables}
\usage{
addNewContinuous(tsid, df, con = NULL)
}
\arguments{
\item{tsid}{The timeseries_id to which the data will be appended. This is a required parameter.}

\item{df}{A data.frame containing the data. Must have columns named 'datetime' and 'value' at minimum. Other optional columns are 'owner', 'contributor', 'approval', 'grade', 'qualifier', 'imputed'; see the \code{adjust_} series of functions to see how these are used.}

\item{con}{A connection to the database, created with \code{\link[DBI:dbConnect]{DBI::dbConnect()}} or using the utility function \code{\link[=AquaConnect]{AquaConnect()}}. If left NULL, a connection will be attempted using AquaConnect() and closed afterwards.}
}
\value{
Nothing; data is added to the database silently.
}
\description{
This function can be used to append new contiuous type data directly to the database without downloading it from a remote source. Differs from \code{\link[=getNewContinuous]{getNewContinuous()}} as the later is used to pull data from a remote before append.
}
