% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addNewContinuous.R
\name{addNewContinuous}
\alias{addNewContinuous}
\title{Add new data to the measurements_continuous or measurements_calculated_daily tables}
\usage{
addNewContinuous(tsid, df, target = "continuous", con = NULL)
}
\arguments{
\item{tsid}{The timeseries_id to which the data will be appended. This is a required parameter.}

\item{df}{A data.frame containing the data. Must have columns named 'datetime' OR 'date', and 'value' at minimum. If 'datetime' is present data will be appended to measurements_continuous, otherwise 'date' will be used to append to measurements_calculated_daily. Other optional columns are 'owner', 'contributor', 'share_with', 'approval', 'grade', 'qualifier', 'imputed'; see the \code{adjust_} series of functions to see how these are used.}

\item{target}{One of 'continuous' or 'daily'. Default is 'continuous'. You would only want to append directly to the 'daily' table if adding pre-calculated daily means with the aim of adding higher frequency data to the 'continuous' table later. As an extra check, the data.frame passed in argument 'df' must contain a column named 'datetime' or 'date' to match this parameter.}

\item{con}{A connection to the database, created with \code{\link[DBI:dbConnect]{DBI::dbConnect()}} or using the utility function \code{\link[=AquaConnect]{AquaConnect()}}. If left NULL, a connection will be attempted using AquaConnect() and closed afterwards.}
}
\value{
Nothing; data is added to the database silently.
}
\description{
Add new data to the measurements_continuous or measurements_calculated_daily tables
}
