% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addNewDiscrete.R
\name{addNewDiscrete}
\alias{addNewDiscrete}
\title{Add new discrete sample data to the database}
\usage{
addNewDiscrete(con, sample, results)
}
\arguments{
\item{con}{A connection to the database, created with \code{\link[DBI:dbConnect]{DBI::dbConnect()}} or using the utility function \code{\link[=AquaConnect]{AquaConnect()}}.}

\item{sample}{A data.frame containing the sample metadata for a single discrete sample. Should contain a single row for a single sample.}

\item{results}{A data.frame containing the results corresponding to the sample. Should contain one row per result.}
}
\value{
The database sample_id for the inserted sample.
}
\description{
Appends a discrete sample and its results to the AquaCache database inside a transaction, returning the created sample_id. The format of the data.frames passed as arguments are defined in the \code{details} section.
}
\details{
The 'sample' data.frame must contain the following columns:
\itemize{
\item 'location_id': a numeric specifying the location_id of the data point from table 'locations'.
\item 'media_id': a numeric specifying the media_id of the data point from table 'medias'.
\item 'datetime': a POSIXct datetime object in UTC 0 time zone, specifying the datetime of the data point.
\item 'collection_method': a numeric specifying the collection_method_id of the data point from table 'collection_methods', such as 1 (observation), 27 (water bottle), or 14 (pump).
\item 'sample_type': a numeric specifying the sample_type_id of the data point from table 'sample_types', such as 1 (grab), 2 (composite), or 3 (integrated).
\item 'owner': the owner of the data, as a character string. This should match entries in the 'organizations' table and an error will be thrown if it does not.
\item 'import_source_id': a numeric specifying the import_source_id of the data point from table 'import_sources' (use for tracking purposes).
Optional columns are:
\item 'target_datetime': a POSIXct datetime object in UTC 0 time zone, specifying an artificial datetime for the data point which can be used for data analysis or plotting purposes.
\item 'note': a character string with a note about the data point(s).
\item 'contributor' the name of the person or organization that contributed the data, as a character string. This should match entries in the 'organizations' table and an error will be thrown if it does not.
\item 'approval': the approval status of the data, as a character string. This should match entries in the 'approvals' table and an error will be thrown if it does not.
\item 'grade': the grade of the data, as a character string. This should match entries in the 'grades' table and an error will be thrown if it does not.
\item 'qualifier': the qualifier of the data, as a character string. This should match entries in the 'qualifiers' table and an error will be thrown if it does not.
}

The 'results' data.frame should contain one row per result and must contain the following columns:
\itemize{
\item 'parameter_id': a numeric specifying the parameter_id of the data point from table 'parameters'.
\item 'result': a numeric specifying the sample's results, matched to the parameters
\item 'result_type': a numeric specifying the result_type_id of the data point from table 'result_types', such as 1 (concentration), 2 (load), or 3 (other).
Additionally, the following columns may need to be included:
\item 'result_condition': a numeric specifying the result condition of the data point from table 'result_conditions', such as "< DL" or "> DL". Only necessary if there are NA values in the 'result' column that should be interpreted as a specific condition. If not provided, rows with NA values will be dropped.
\item 'result_condition_value': a numeric specifying the value of the result condition, such as 0.1 for "< DL 0.1". Necessary if column 'result_condition' is provided AND contains values of 1 or 2, i.e. 'Below Detection/Quantification Limit' or 'Above Detection/Quantification Limit'.
\item 'sample_fraction_id': a numeric specifying the sample_fraction_id of the data point from table 'sample_fractions', such as 19 ('total'), 5 ('dissolved'), or 18 ('suspended'). Required if the column 'sample_fraction' in table 'parameters' is TRUE for the parameter in question.
\item 'result_speciation_id': a numeric specifying the result_speciation_id of the data point from table 'result_speciations', such as 3 (as CaCO3), 5 (as CN), or 44 (of S). Required if the column 'result_speciation' in table 'parameters' is TRUE for the parameter in question.
}
}
