% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_functions.R
\name{adjust_data_sharing_agreement}
\alias{adjust_data_sharing_agreement}
\title{Adjust the data sharing agreement of a timeseries in the database}
\usage{
adjust_data_sharing_agreement(con, timeseries_id, data, delete = FALSE)
}
\arguments{
\item{con}{A connection to the database with write privileges to the 'timeseries_data_sharing_agreements' table.}

\item{timeseries_id}{The target timeseries_id}

\item{data}{A data.frame with columns for 'datetime' and 'data_sharing_agreement_id'. 'datetime' should be POSIXct and 'data_sharing_agreement_id' should refer to column 'document_id' of table 'files.documents'.}

\item{delete}{Logical. If TRUE, the function will delete data sharing agreements which come entirely after the start of 'data'. This ensures synchronization with remote data stores and is called from the 'synchronize' functions.}
}
\value{
Modifies the 'timeseries_data_sharing_agreements' table in the database.
}
\description{
Adjust the data sharing agreement of a timeseries in the database
}
