% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_functions.R
\name{adjust_grade}
\alias{adjust_grade}
\title{Adjust the grade of a timeseries in the database}
\usage{
adjust_grade(con, timeseries_id, data)
}
\arguments{
\item{con}{A connection to the database with write privileges to the 'approvals' and 'measurements_continuous' tables.}

\item{timeseries_id}{The target timeseries_id}

\item{data}{A data.frame with columns for 'datetime' and 'approval'. 'datetime' should be POSIXct and 'approval' should either character (in which case it must refer to entries in column 'approval_type_code' of table 'approval_types' or integer/numeric, in which case it must refer to column 'approval_type_id' of the same table.}
}
\value{
Modifies the 'approvals' table in the database.
}
\description{
Adjust the grade of a timeseries in the database
}
