% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_functions.R
\name{adjust_owner}
\alias{adjust_owner}
\title{Adjust the owner of a timeseries in the database}
\usage{
adjust_owner(con, timeseries_id, data)
}
\arguments{
\item{con}{A connection to the database with write privileges to the 'owners' and 'measurements_continuous' tables.}

\item{timeseries_id}{The target timeseries_id}

\item{data}{A data.frame with columns for 'datetime' and 'owner'. 'datetime' should be POSIXct and 'owner' should be either character (in which case it must refer to entries in column 'name' of table 'owners_contributors_operators' or integer/numeric, in which case it must refer to column 'owner_contributor_id' of the same table.}
}
\value{
Modifies the 'owners' table in the database.
}
\description{
Adjust the owner of a timeseries in the database
}
