% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_functions.R
\name{adjust_qualifier}
\alias{adjust_qualifier}
\title{Adjust the qualifier of a timeseries in the database}
\usage{
adjust_qualifier(con, timeseries_id, data)
}
\arguments{
\item{con}{A connection to the database with write privileges to the 'qualifiers' and 'measurements_continuous' tables.}

\item{timeseries_id}{The target timeseries_id}

\item{data}{A data.frame with columns for 'datetime' and 'qualifier'. 'datetime' should be POSIXct and 'qualifier' should either character (in which case it must refer to entries in column 'qualifier_type_code' of table 'qualifiers' or integer/numeric, in which case it must refer to column 'qualifier_type_id' of the same table.}
}
\value{
Modifies the 'qualifiers' table in the database.
}
\description{
Adjust the qualifier of a timeseries in the database
}
