% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AquaConnect.R
\name{AquaConnect}
\alias{AquaConnect}
\title{Connect to the AquaCache database}
\usage{
AquaConnect(
  name = "AquaCache",
  host = Sys.getenv("AquaCacheHost"),
  port = Sys.getenv("AquaCachePort"),
  username = Sys.getenv("AquaCacheAdminUser"),
  password = Sys.getenv("AquaCacheAdminPass"),
  silent = FALSE,
  dev = FALSE
)
}
\arguments{
\item{name}{Database name.}

\item{host}{Database host address. By default searches the .Renviron file for parameter:value pair of form AquaCacheHost:"hostname".}

\item{port}{Connection port. By default searches the .Renviron file for parameter:value pair of form AquaCachePort:"1234".}

\item{username}{Username. By default searches the .Renviron file for parameter:value pair of form AquaCacheAdminUser:"username".}

\item{password}{Password. By default searches the .Renviron file for parameter:value pair of form AquaCacheAdminPass:"password".}

\item{silent}{TRUE suppresses messages except for errors.}

\item{dev}{TRUE appends "_dev" to the database name.}
}
\value{
A connection to the database.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function exists to facilitate connecting to the hydrology database. A nearly identical function exists in package YGwater, but this one by default uses admin privileges while the other uses read-only privileges. Database superusers will have access to all database records, but other users will be asked to provide their username and password to access records other than the 'public' ones. Note that this is \emph{not necessarily} the same username and password as the one used to log into the database itself.
}
