% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AquaConnect.R
\name{AquaConnect}
\alias{AquaConnect}
\title{Connect to the aquacache database}
\usage{
AquaConnect(
  name = Sys.getenv("aquacacheName"),
  host = Sys.getenv("aquacacheHost"),
  port = Sys.getenv("aquacachePort"),
  username = Sys.getenv("aquacacheAdminUser"),
  password = Sys.getenv("aquacacheAdminPass"),
  silent = FALSE
)
}
\arguments{
\item{name}{Database name. By default searches the .Renviron file for parameter=value pair of form aquacacheName="name". If you want to connect to a different database, pass the name of that database here.}

\item{host}{Database host address. By default searches the .Renviron file for parameter:value pair of form aquacacheHost:"hostname".}

\item{port}{Connection port. By default searches the .Renviron file for parameter:value pair of form aquacachePort:"1234".}

\item{username}{Username. By default searches the .Renviron file for parameter:value pair of form aquacacheAdminUser:"username".}

\item{password}{Password. By default searches the .Renviron file for parameter:value pair of form aquacacheAdminPass:"password".}

\item{silent}{TRUE suppresses messages except for errors and login messages.}
}
\value{
A connection to the database.
}
\description{
This function exists to facilitate connecting to the hydrology database. A nearly identical function exists in package YGwater, but this one by default uses admin privileges while the other uses read-only privileges.

See Details for more information for developers.
}
\details{
A check is performed to see if there are any patches to apply to the database on every connection. If there are, the user is prompted to apply them. It is highly recommended that you do so before running any other functions from this package.
}
