% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_stats.R
\name{calculate_stats}
\alias{calculate_stats}
\title{Calculate daily means and statistics}
\usage{
calculate_stats(locations = NULL, path = NULL)
}
\arguments{
\item{locations}{A data.frame containing, at a minimum, columns named 'location', and 'parameter', necessary to identify the exact records in need of updating.}

\item{path}{The path to the hydrometric database, passed to WRBtools::hydroConnect.}
}
\value{
Updated entries in the 'daily' table
}
\description{
Calculate daily means and statistics
}
