% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_stats.R
\name{calculate_stats}
\alias{calculate_stats}
\title{Calculate daily means and statistics}
\usage{
calculate_stats(con = NULL, timeseries_id, start_recalc = NULL)
}
\arguments{
\item{con}{A connection to the database. If NULL, a connection will be created and closed by the function.}

\item{timeseries_id}{The timeseries_ids you wish to have updated, as character or numeric vector. Specifying 'all' will work on all timeseries.}

\item{start_recalc}{The day on which to start daily calculations, as a vector of one element OR as NULL. If NULL will recalculate the last two days in case there is new realtime data, plus two days in the past to account for possible past calculations with incomplete data.}

\item{day}{A Date object.}

\item{datetime}{A POSIXct or POSIXlt object.}

\item{offset}{The timezone offset in hours from UTC.}
}
\value{
Updated entries in the 'measurements_calculated_daily' table.
}
\description{
Calculates daily means from data in the measurements_continuous table as well as derived statistics for each day (historical min, max, q10, q25, q50 (mean), q75, q90), using the timezone specified in table 'timeseries' to define the start/end of days. Derived daily statistics are for each day of year \strong{prior} to the current date for historical context, with the exception of the first day of record for which only the min/max are populated with the day's value. Any data graded as 'unusable' is excluded from calculations. February 29 calculations are handled differently: see details.

Water Survey of Canada daily means are dealt with differently than other timeseries: daily means provided in the HYDAT database take precedence over means calculated from data in table 'measurements_continuous'.

Some continuous measurement data may have a period of greater than 1 day. In these cases it would be impossible to calculate daily statistics, so this function explicitly excludes data points with a period greater than P1D.

This function is meant to be called from within \code{\link[=dailyUpdate]{dailyUpdate()}}, but is exported in case a need arises to calculate daily means and statistics in isolation or in another function. It \emph{must} be used with a database created by this package, or one with identical tables.
}
\details{
Calculating daily statistics for February 29 is complicated: due to a paucity of data, this day's statistics are liable to be very mismatched from those of the preceding and succeeding days if calculated based only on Feb 29 data. Consequently, statistics for these days are computed by averaging those of Feb 28 and March 1, ensuring a smooth line when graphing mean/min/max/quantile parameters. This necessitates waiting for complete March 1st data, so Feb 29 means and stats will be delayed until March 2nd.
}
