% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_stats.R
\name{calculate_stats}
\alias{calculate_stats}
\title{Calculate daily means and statistics}
\usage{
calculate_stats(timeseries = NULL, path = NULL)
}
\arguments{
\item{timeseries}{A data.frame containing, at a minimum, columns named 'location', and 'parameter', necessary to identify the exact records in need of updating.}

\item{path}{The path to the hydrometric database, passed to WRBtools::hydroConnect.}
}
\value{
Updated entries in the 'daily' table
}
\description{
This function is meant to be called from within hydro_update_daily, but is exported just in case a need arises to calculate daily means and statistics in isolation. It \emph{must} be used with a database created by this package, or one with identical table and column names.
}
