% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_DB.R
\name{create_test_db}
\alias{create_test_db}
\title{Create a small test database}
\usage{
create_test_db(
  name = "aquacache",
  host = Sys.getenv("aquacacheHost"),
  port = Sys.getenv("aquacachePort"),
  username = Sys.getenv("aquacacheAdminUser"),
  password = Sys.getenv("aquacacheAdminPass"),
  outpath = testthat::test_path("fixtures"),
  replace = TRUE,
  pg_dump = NULL,
  psql = NULL,
  continuous_locations = NULL,
  discrete_locations = NULL,
  start_datetime = "2015-12-31 00:00",
  end_datetime = "2023-01-02 00:00",
  delete = TRUE
)
}
\arguments{
\item{name}{Target database name (i.e. the one to be dumped). By default, it is set to "aquacache". If you want to dump a different database, specify its name here.}

\item{host}{Database host address. By default searches the .Renviron file for parameter:value pair of form aquacacheHost="hostname".}

\item{port}{Connection port. By default searches the .Renviron file for parameter:value pair of form aquacachePort="1234".}

\item{username}{Username. By default searches the .Renviron file for parameter:value pair of form aquacacheAdminUser="username".}

\item{password}{Password. By default searches the .Renviron file for parameter:value pair of form aquacacheAdminPass="password".}

\item{outpath}{The path to the folder/location where the .sql dump will be saved. If "choose", the user will be prompted to select a folder. Default is this package's tests/testthat/fixtures folder.}

\item{replace}{If TRUE and a file of name testdb.sql.gz exists in \code{outpath} will attempt to replaced the file.}

\item{pg_dump}{The path to the pg_dump utility. By default (NULL), the function searches the PATH for the utility, but this might not always work.}

\item{psql}{The path to the psql utility. By default (NULL), the function searches the PATH for the utility, but this might not always work.}

\item{continuous_locations}{Character vector of location codes to retain in continuous tables. Defaults to the first location found in the database when \code{NULL}.}

\item{discrete_locations}{Character vector of location codes to retain in discrete tables.  Defaults to \code{continuous_locations} when \code{NULL}.}

\item{start_datetime}{The start datetime for the data to be copied.  If \code{NULL}, data is copied from the beginning of records to \code{end_datetime} for the affected locations. This does not apply to the 'measurements_calculated_daily' table, which is always copied from the beginning of records.}

\item{end_datetime}{The end datetime for the data to be copied.  If \code{NULL}, all data is copied from \code{start_datetime} to the end of records for the affected locations. This \strong{does} apply to the 'measurements_calculated_daily' table.}

\item{delete}{If TRUE, will delete the test_temp database after the function is done. If FALSE, the database will remain for further testing.}
}
\value{
An SQL file containing the schema definition saved to the 'outfile' path.
}
\description{
This function uses the \code{pg_dump} utility to create a dump of the schema and a subset of the data in an aquacache PostgreSQL database. The schema dump is saved to an SQL file in the specified output path. The resulting SQL includes a command to set the database \code{search_path} on restore so that queries without schema qualifiers work. You must call this function from a machine with the PostgreSQL pg_dump utility installed.
}
