% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dailyUpdate.R
\name{dailyUpdate}
\alias{dailyUpdate}
\title{Daily update of hydro database}
\usage{
dailyUpdate(con = hydrometConnect(silent = TRUE), timeseries_id = "all")
}
\arguments{
\item{con}{A connection to the database, created with \code{\link[DBI:dbConnect]{DBI::dbConnect()}} or using the utility function \code{\link[=hydrometConnect]{hydrometConnect()}}.}

\item{timeseries_id}{The timeseries_ids you wish to have updated, as character or numeric vector. Defaults to "all".}
}
\value{
The database is updated in-place, and diagnostic messages are printed to the console.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function is intended to be run on a daily or near-daily basis to ensure data integrity. Pulls in new data, calculates statistics where necessary, and performs cross-checks on several tables (see details for more information).
}
\details{
Calls several functions in sequence: \code{\link[=getNewContinuous]{getNewContinuous()}} to pull in new data into the measurements_continuous table, \code{\link[=getNewDiscrete]{getNewDiscrete()}} to pull in new data to the measurements_discrete table, \code{\link[=getNewImages]{getNewImages()}} to get new images in a series, \code{\link[=getNewRasters]{getNewRasters()}} to get new model rasters, \code{\link[=update_hydat]{update_hydat()}} to check for a new HYDAT database version (hydrometric data from the WSC) and incorporate daily means which differ from those already calculated in the dabatabase, \code{\link[=update_hydat_datums]{update_hydat_datums()}} to update the datums table with any new datums present in HYDAT, and \code{\link[=calculate_stats]{calculate_stats()}} to calculate new statistics where necessary.

Note that new timeseries should be added using function \code{\link[=addHydrometTimeseries]{addHydrometTimeseries()}}.

Any timeseries labelled as downloadAquarius in the source_fx column in the timeseries table will need your Aquarius username, password, and server URL present in your .Renviron profile, or those three parameters entered in the column source_fx_args: see downloadAquarius for more information about that function, and \code{\link[=addHydrometTimeseriesTemplate]{addHydrometTimeseriesTemplate()}} for details on how to format the parameters to pass to \code{\link[=addHydrometTimeseries]{addHydrometTimeseries()}}.
}
