% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dbTransBegin}
\alias{dbTransBegin}
\title{Begin a transaction}
\usage{
dbTransBegin(con, silent = TRUE)
}
\arguments{
\item{con}{A database connection object.}
}
\value{
A boolean indicating whether a transaction was started
}
\description{
Begin a transaction on a database connection. This function creates a temporary table to ensure the transaction is active.
}
