% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadAquarius.R
\name{downloadAquarius}
\alias{downloadAquarius}
\title{Get formatted timeseries data from Aquarius.}
\usage{
downloadAquarius(
  location,
  parameter,
  start_datetime,
  end_datetime = Sys.Date(),
  login = Sys.getenv(c("AQUSER", "AQPASS")),
  server = Sys.getenv("AQSERVER"),
  con = NULL
)
}
\arguments{
\item{location}{The location ID, exactly as visible in Aquarius web portal, as a character vector of length 1. Typically of form \verb{29EA001} or \code{YOWN-0804}.}

\item{parameter}{The timeseries name, exactly as visible in Aquarius web portal, as a character vector of length 1. Typically of form \code{Wlevel_bgs.Calculated}.}

\item{start_datetime}{The first day or instant for which you want information. You can specify a Date object, POSIXct object, or character vector of form yyyy-mm-dd or yyyy-mm-dd HH:mm:ss. If specifying a POSIXct the UTC offset associated with the time will be used, otherwise UTC 0 will be assumed. If only a date is specified it will be assigned the first moment of the day. Times requested prior to the actual timeseries start will be adjusted to match available data.}

\item{end_datetime}{The last day or instant for which you want information. You can specify a Date object, POSIXct object, or character vector of form yyyy-mm-dd or yyyy-mm-dd HH:mm:ss. If specifying a POSIXct the UTC offset associated with the time will be used, otherwise UTC 0 will be assumed. If only a date is specified it will be assigned the last moment of the day. Times requested prior to the actual timeseries end will be adjusted to match available data.}

\item{login}{Your Aquarius login credentials as a character vector of two. Default pulls information from your .renviron file; see details.}

\item{server}{The URL for your organization's Aquarius web server. Default pulls from your .renviron file; see details.}

\item{con}{A connection to the aquacache database, necessary to allow for the mapping of Aquarius approvals, grades, and qualifiers to the database. If left NULL connection will be made and closed automatically.}
}
\value{
A data.frame of timeseries data with columns for datetime, value, grade, approval.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Pared-down and modified version of YGWater::aq_download.
\subsection{Passwords and server credentials:}{

To store login credentials in your .renviron file call \code{\link[usethis:edit]{usethis::edit_r_environ()}} and enter your username and password as value pairs, as AQUSER="your username" and AQPASS="your password". The server should be entered at server="your_server_url". You can also store credentials in the timeseries table in the column source_fx_args, but beware that these credentials are then sitting in the database un-encrypted.
}

\subsection{Grades, approvals, qualifiers:}{

This function will attempt to map Aquarius grades, approvals, and qualifiers to the database defaults. The function was designed around the Yukon Water Resources Branch Aquarius schemes, but 'forks' can be implemented for each based on the server URL. If you need to implement a fork, see the commented-out example in the function at line 116.
}
}
