% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadECCCwq.R
\name{downloadECCCwq}
\alias{downloadECCCwq}
\title{Bring ECCC water quality data to AquaCache}
\usage{
downloadECCCwq(
  location,
  file,
  key = system.file("import_keys", "ECCCwq.csv", package = "AquaCache"),
  start_datetime,
  end_datetime = Sys.time(),
  con = NULL
)
}
\arguments{
\item{location}{The location code associated with the ECCC monitoring site. Must be a valid location code in the \code{SITE_NO} field of the ECCC water quality .csv files.}

\item{file}{Path (URL) to the ECCC water quality .csv file containing the data to be imported for the specified location.}

\item{key}{Path to the import key .csv file defining the parameter mappings for ECCC water quality data.}

\item{start_datetime}{Start datetime (inclusive) from which to fetch measurements. Specify as class Date, POSIXct OR as character string which can be interpreted as POSIXct. If character, UTC offset of 0 will be assigned, otherwise conversion to UTC 0 will be performed on POSIXct class input. If date, time will default to 00:00 to capture whole day.}

\item{end_datetime}{End datetime (inclusive) to which to fetch measurements. Specify as class Date, POSIXct OR as character string which can be interpreted as POSIXct. If character, UTC offset of 0 will be assigned, otherwise conversion to UTC 0 will be performed on POSIXct class input. If Date, time will default to 23:59:59 to capture whole day.}

\item{con}{A connection to the aquacache database, only used if an offset is calculated for an old_loc. If not provided, a connection will be attempted using AquaConnect().}
}
\value{
A data.frame object with the requested data. If there are no new data points the data.frame will have 0 rows.
}
\description{
Brings in water quality data from ECCC long-term monitoring sites and transforms them into the aquacache database format.
}
