% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadECCCwx.R
\name{downloadECCCwx}
\alias{downloadECCCwx}
\title{Get realtime data from ECCC met locations (observations)}
\usage{
downloadECCCwx(
  location,
  parameter,
  start_datetime,
  end_datetime = Sys.time(),
  interval,
  con = NULL
)
}
\arguments{
\item{location}{An ECCC Station ID (not to be mistaken for other IDs such as the Nav Canada ID, the WMO ID, or the Climate ID). See \code{\link[weathercan:stations]{weathercan::stations()}} for help finding the right ID.}

\item{parameter}{The name of the column containing the desired data, as output in the data.frame given by \code{\link[weathercan:weather_dl]{weathercan::weather_dl()}}. Taken from the source_fx_args in the timeseries table table. Note that this column name varies depending on the interval specified (hour, day, month).}

\item{start_datetime}{Specify as class Date, POSIXct OR as character string which can be interpreted as POSIXct. If character, UTC offset of 0 will be assigned, otherwise conversion to UTC 0 will be performed on POSIXct class input. If date, time will default to 00:00 to capture whole day.}

\item{end_datetime}{Specify as class Date, POSIXct OR as character string which can be interpreted as POSIXct. If character, UTC offset of 0 will be assigned, otherwise conversion to UTC 0 will be performed on POSIXct class input. If Date, time will default to 23:59:59 to capture whole day.}

\item{interval}{The interval to pass to \code{\link[weathercan:weather_dl]{weathercan::weather_dl()}}, one of "hour", "day", "month".}

\item{con}{A connection to the aquacache database, necessary to allow for the mapping of Aquarius approvals, grades, and qualifiers to the database. If left NULL connection will be made and closed automatically.}
}
\value{
A data.frame of hydrometric data, with datetimes in UTC-0.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

A function used to fetch weather data from ECCC, using the weathercan package for speed and simplicity. Since ECCC weather data comes in as a tibble with ~36 rows (all parameters)and there is no way to tailor the request to a single parameter, this function will save the output of the first download as an .rdata file to the session temporary folder. Subsequent runs of the function will search the temporary folder for a suitable file and attempt to use it, downloading again only if no suitable file is found. Temporary folder contents are deleted when the R session is closed.
}
