% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadECCCwxMinute.R
\name{downloadECCCwxMinute}
\alias{downloadECCCwxMinute}
\title{Get minute-resolution weather observations from the ECCC SWOB realtime API}
\usage{
downloadECCCwxMinute(
  location,
  parameter,
  start_datetime,
  end_datetime = Sys.time(),
  con = NULL
)
}
\arguments{
\item{location}{A four-letter station code used by the SWOB realtime API,
such as \code{"CVXY"}. See details.}

\item{parameter}{The SWOB element name to extract, or a supported alias from
\code{\link[=downloadECCCwx]{downloadECCCwx()}} such as \code{"temp"}, \code{"wind_spd"}, \code{"wind_dir"},
\code{"wind_gust"}, \code{"stn_press"}, or \code{"dew_point"}. See details.}

\item{start_datetime}{Specify as class Date, POSIXct OR as character string
which can be interpreted as POSIXct. If character, UTC offset of 0 will be
assigned, otherwise conversion to UTC 0 will be performed on POSIXct class
input. If date, time will default to 00:00 to capture whole day.}

\item{end_datetime}{Specify as class Date, POSIXct OR as character string
which can be interpreted as POSIXct. If character, UTC offset of 0 will be
assigned, otherwise conversion to UTC 0 will be performed on POSIXct class
input. If Date, time will default to 23:59:59 to capture whole day.}

\item{con}{A connection to the aquacache database, necessary to allow for the
mapping of approvals, grades, qualifiers, and owners to the database. If
left NULL connection will be made and closed automatically.}
}
\value{
A data.table object of weather observations, with datetimes in UTC-0.
}
\description{
Fetches minute-resolution SWOB observations from the filtered realtime API at
https://api.weather.gc.ca/. The query is restricted to one station, one
parameter, and the requested datetime interval, then returned as CSV for fast
parsing into AquaCache's standard one-timeseries-at-a-time workflow.
}
\details{
Raw parameter names are available from the SWOB queryables page at
https://api.weather.gc.ca/collections/swob-realtime/queryables?f=html. The
function also accepts familiar aliases such as \code{"temp"}, \code{"wind_spd"},
\code{"wind_dir"}, \code{"wind_gust"}, \code{"stn_press"}, and \code{"dew_point"}, though you are
recommended to use exact SWOB parameter names when possible.
}
