% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadEQWin.R
\name{downloadEQWin}
\alias{downloadEQWin}
\title{Bring EQWin water quality data into the aquacache database}
\usage{
downloadEQWin(
  location,
  start_datetime,
  end_datetime = Sys.time(),
  EQpath = NULL,
  key = NULL
)
}
\arguments{
\item{location}{The location code associated with the EQWin station. Must match an entry in the eqstns table, column StnCode.}

\item{start_datetime}{Specify as class Date, POSIXct OR as character string which can be interpreted as POSIXct. If character, UTC offset of 0 will be assigned, but if POSIXct the object's tzone attribute will be preserved. If date, time will default to 00:00 to capture whole day.}

\item{end_datetime}{Same as start_datetime but not quite!}

\item{EQpath}{The file path to the target EQWin database.}

\item{key}{A path to the .csv file mapping EQWin parameter codes to aquacache parameter codes.}
}
\value{
A data.frame object with the requested data. If there are no new data points the data.frame will have 0 rows.
}
\description{
Brings in water quality data from EQWin databases.
}
