% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadERA5.R
\name{downloadERA5}
\alias{downloadERA5}
\title{Get ERA5 rasters}
\usage{
downloadERA5(start_datetime, clip = "YT", param, user, key)
}
\arguments{
\item{start_datetime}{The datetime from which to start looking for new rasters}

\item{clip}{The two-digit abbreviation(s) as per \href{https://www12.statcan.gc.ca/census-recensement/2021/ref/dict/tab/index-eng.cfm?ID=t1_8}{Canadian Census} for the province(s) with which to clip the HRDPA. A 300 km buffer is added beyond the provincial boundaries. Set to NULL for no clip.}

\item{param}{The parameter for which to get new rasters. Currently only "APCP_Sfc" is supported.}

\item{user}{The username to use for ECMWF authentication.}

\item{key}{The key to use for EMCWF authentication.}
}
\value{
A list of lists, where each element consists of the target raster as well as associated attributes.
}
\description{
Get ERA5 rasters
}
