% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadHRDPA.R
\name{downloadHRDPA}
\alias{downloadHRDPA}
\title{Get HRDPA rasters}
\usage{
downloadHRDPA(
  start_datetime,
  url = "https://dd.weather.gc.ca/model_hrdpa/2.5km/",
  clip = NULL
)
}
\arguments{
\item{start_datetime}{The datetime from which to start looking for new rasters}

\item{url}{The url from which to get new rasters.}

\item{clip}{The two-digit abbreviation(s) as per \href{https://www12.statcan.gc.ca/census-recensement/2021/ref/dict/tab/index-eng.cfm?ID=t1_8}{Canadian Census} for the province(s) with which to clip the HRDPA. A 300 km buffer is added beyond the provincial boundaries. Set to NULL for no clip.}
}
\value{
A list of lists, where each element consists of the target raster as well as associated attributes.
}
\description{
Get HRDPA rasters
}
