% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadNupointImages.R
\name{downloadNupointImages}
\alias{downloadNupointImages}
\title{Get images from NuPoint SFTP server}
\usage{
downloadNupointImages(
  location,
  start_datetime,
  username = Sys.getenv("nupointUser"),
  password = Sys.getenv("nupointPass"),
  url = Sys.getenv("nupointServer"),
  port = Sys.getenv("nupointPort"),
  folder = Sys.getenv("nupointFolder"),
  save_path = NULL,
  delete = TRUE
)
}
\arguments{
\item{location}{The location for which to get images.}

\item{start_datetime}{The earliest datetime to start pulling images from.}

\item{username}{Username to use for password-protected login.}

\item{password}{Password to use for protected login.}

\item{url}{The URL or IP address from which to get new images.}

\item{port}{The port on which to connect to the SFTP server.}

\item{folder}{The folder in which to look for new images.}

\item{save_path}{Optional; path in which to save the image.}

\item{delete}{Should the files be deleted from the SFTP site once they've been successfully been fetched?}
}
\value{
A list object of type 'response' containing the image (object$content) and metadata regarding the object and how/when it was fetched.
}
\description{
Fetches images from a NuPoint SFTP server. Default URL/IP address is Yukon-specific. To fetch from the WSC instead see \code{\link[=downloadWSCImages]{downloadWSCImages()}}. Intended to be called by function \code{\link[=getNewImages]{getNewImages()}}.
}
