% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EQConnect.R
\name{EQConnect}
\alias{EQConnect}
\title{Connect to the EQWin database}
\usage{
EQConnect(
  path = "//carver/infosys/EQWin/WR/DB/Water Resources.mdb",
  silent = FALSE
)
}
\arguments{
\item{path}{Full path to the database including extension.}

\item{silent}{TRUE suppresses messages except for errors.}
}
\value{
A connection to the database.
}
\description{
This function exists to facilitate connecting to the EQWin Access database. Requires installation of Microsoft Access redistributable of the right bitness for the version of R you're running. See https://www.microsoft.com/en-us/download/details.aspx?id=54920 for installation. Note that 64 bit R needs a 64 bit installation.
}
