% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchFile.R
\name{fetchDocument}
\alias{fetchDocument}
\title{Extract document from postgreSQL DB}
\usage{
fetchDocument(
  id,
  con = hydrometConnect(silent = TRUE),
  save_dir = NULL,
  save_name = NULL
)
}
\arguments{
\item{id}{The ID number from column 'document_id' of table 'documents.}

\item{con}{A connection to the database.}

\item{save_dir}{A directory in which to write the file.}

\item{save_name}{The name to give the file, \emph{without} extension.}
}
\value{
A data.frame containing the row(s) identified by the specified ID. One column will contain the binary object. If asked for, the document only will be saved to file.
}
\description{
Extracts a document stored in the hydromet database as BYTEA type. For extracting images see \code{\link[=fetchImage]{fetchImage()}}, and for extracting rasters to R (which automatically sets the CRS) see \code{\link[rpostgis:pgGetRast]{rpostgis::pgGetRast()}}. Calls function internal package function fetchFile.
}
