% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNewContinuous.R
\name{getNewContinuous}
\alias{getNewContinuous}
\title{Get new continuous-category data}
\usage{
getNewContinuous(
  con = NULL,
  timeseries_id = "all",
  active = "default",
  dbName = NULL,
  dbHost = NULL,
  dbPort = NULL,
  dbUser = NULL,
  dbPass = NULL,
  verbose = FALSE
)
}
\arguments{
\item{con}{A connection to the database, created with \code{\link[DBI:dbConnect]{DBI::dbConnect()}} or using the utility function \code{\link[=AquaConnect]{AquaConnect()}}. NULL will create a connection and close it afterwards, otherwise it's up to you to close it after. If you wish to run this function in parallel you MUST leave this argument NULL. If you also specify connection parameters in later arguments they will be used, otherwise the function will use the AquaConnect defaults.}

\item{timeseries_id}{The timeseries_ids you wish to have updated, as character or numeric vector. Defaults to "all", which means all timeseries of category 'continuous'.}

\item{active}{Sets behavior for import of new data. If set to 'default', the function will look to the column 'active' in the 'timeseries' table to determine if new data should be fetched. If set to 'all', the function will ignore the 'active' column and import all data.}

\item{dbName}{The name of the database to connect to. If left NULL, the function will use the default database name from the .Renviron file as per \code{\link[=AquaConnect]{AquaConnect()}}.}

\item{dbHost}{The host address of the database. If left NULL, the function will use the default host address from the .Renviron file as per \code{\link[=AquaConnect]{AquaConnect()}}.}

\item{dbPort}{The port of the database. If left NULL, the function will use the default port from the .Renviron file as per \code{\link[=AquaConnect]{AquaConnect()}}.}

\item{dbUser}{The username for the database. If left NULL, the function will use the default username from the .Renviron file as per \code{\link[=AquaConnect]{AquaConnect()}}.}

\item{dbPass}{The password for the database. If left NULL, the function will use the default password from the .Renviron file as per \code{\link[=AquaConnect]{AquaConnect()}}.}

\item{verbose}{If TRUE, will print the timeseries_id of each iteration as it is processed. Default is FALSE.}
}
\value{
The database is updated in-place, and a data.frame is generated with one row per updated location.
}
\description{
Retrieves new real-time data starting from the last local data point. The
active assignment with the lowest fetch priority in
\code{continuous.timeseries_source_adapters} supplies each timeseries' source
function and arguments. Timeseries without an eligible assignment are
ignored. See \code{\link[=addACTimeseries]{addACTimeseries()}} for creating assignments.

If you are a developer, note that download or source functions MUST be registered in AquaCache using function \code{\link[=registerSourceAdapterArguments]{registerSourceAdapterArguments()}}, and that this operation would normally be completed using the 'patch' system. See patch_56.R for examples.
\subsection{Default arguments passed to 'source_fx' functions:}{

This function passes \code{start_datetime}, defaulting to the instant after the
last local point. Other fetch parameters come from the selected assignment's
\code{source_fx_args} JSON object.
}

\subsection{Multi-timeseries transmission sources:}{

Source-adapter capabilities registered by database Patch 56 control
automatic argument injection and cache-sharing groups. For example,
\code{downloadNESDIS} receives the current \code{timeseries_id} automatically and
timeseries sharing a DCP address reuse one raw transmission payload.
Platform identifiers and field mappings remain database metadata; they do
not need to be duplicated in each timeseries assignment's arguments.

Every assigned \code{source_fx} must have an enabled entry in
\code{public.source_adapter_capabilities} for the continuous domain. This keeps
scheduled imports and administrative clients aligned on the supported
continuous-data adapters.
}

\subsection{Assigning measurement periods:}{

With the exception of "instantaneous" timeseries which automatically receive a period of "00:00:00" (0 time), the period associated with measurements (ex: 1 hour precipitation sum) is derived from the interval between measurements UNLESS a period column is provided by the source function (column source_fx, may also depend on source_fx_args). This function typically fetches only a few hours of measurements at a time, so if the interval cannot be conclusively determined from the new data (i.e. hourly measurements over four hours with two measurements missed) then additional data points will be pulled from the database.

If a period supplied by any data fetch function cannot be coerced to an period object acceptable to "duration" data type, NULL values will be entered to differentiate from instantaneous periods of "00:00:00".
}

\subsection{Sharing privileges and ownership}{

This is dictated by the timeseries table, and checked prior to passing data through view tables to public users.
}
}
