% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNewDiscrete.R
\name{getNewDiscrete}
\alias{getNewDiscrete}
\title{Get new discrete-type data}
\usage{
getNewDiscrete(con = hydrometConnect(silent = TRUE), timeseries_id = "all")
}
\arguments{
\item{con}{A connection to the database, created with \code{\link[DBI:dbConnect]{DBI::dbConnect()}} or using the utility function \code{\link[=hydrometConnect]{hydrometConnect()}}.}

\item{timeseries_id}{The timeseries_ids you wish to have updated, as character or numeric vector. Defaults to "all", which means all timeseries of category 'discrete'.}
}
\value{
The database is updated in-place, and a data.frame is generated with one row per updated location.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Retrieves new discrete data starting from the last data point in the local database, using the function specified in the timeseries table column "source_fx". Only works on stations that are ALREADY in the discrete table and that have a proper entry in the timeseries table; refer to \code{\link[=addHydrometTimeseries]{addHydrometTimeseries()}} for how to add new stations. Does not work on any timeseries of category "continuous": for that, use \code{\link[=getNewContinuous]{getNewContinuous()}}. Timeseries with no specified souce_fx will be ignored.
\subsection{Default arguments passed to 'source_fx' functions:}{

This function passes default arguments to the "source_fx" function: 'location' gets the location as entered in the 'timeseries' table, 'param_code' gets the parameter code defined in the 'settings' table, and start_datetime defaults to the instant after the last point already existing in the DB. Each of these can however be set using the "source_fx_args" column in the "timeseries" table; refer to \code{\link[=addHydrometTimeseries]{addHydrometTimeseries()}} for a description of how to formulate these arguments.
}
}
