% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNewRasters.R
\name{getNewRasters}
\alias{getNewRasters}
\title{Get new rasters}
\usage{
getNewRasters(
  raster_series_ids = "all",
  con = NULL,
  keep_forecasts = "selective",
  active = "default",
  start_datetime = NULL,
  end_datetime = NULL,
  replace = TRUE
)
}
\arguments{
\item{raster_series_ids}{A vector of raster_series_id's. Default 'all' fetches all ids in the raster_series_index table.}

\item{con}{A connection to the database. Default is NULL, which will use the package default connection settings and close the connection afterwards.}

\item{keep_forecasts}{Should forecasts be kept or replaced? Default is 'selective', which keeps only rasters for which there is no new forecast. 'all' keeps all forecasts, and 'none' replaces all forecasts. This does not apply to raster series labelled as 'reanalysis'}

\item{active}{Sets behavior for import of new rasters for raster series. If set to 'default', the column 'active' in the raster_series_index table will determine whether to get new raster or not. If set to 'all', all image series will be fetched regardless of the 'active' column.}

\item{start_datetime}{A start datetime to fetch rasters from. By default, fetches from the last raster end_datetime + 1 second, however this parameter is provided for flexibility. If combined with \code{replace = TRUE}, could be used to replace rasters from a specific datetime to \code{end_datetime}. Specify as POSIXct or something coercible to POSIXct; coercion will be done with to UTC time zone. Only used for reanalysis rasters!}

\item{end_datetime}{An end datetime to fetch rasters to. By default, fetches to the current time, however this parameter is provided for flexibility. If combined with \code{replace = TRUE}, could be used to replace rasters from \code{start_datetime} to a specific datetime. Specify as POSIXct or something coercible to POSIXct; coercion will be done with to UTC time zone. Only used for reanalysis rasters!}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Retrieves new data corresponding to entries in the table "raster_series_index" for which the column 'public' is TRUE. You can add a new raster series with \code{\link[=addACRasterSeries]{addACRasterSeries()}}. As with the timeseries and images table, fetching new data depends on the function listed in the source_fx column of the relevant table and optionally on parameters in column source_fx_args. Refer to \code{\link[=addACTimeseries]{addACTimeseries()}} for a description of how to formulate these arguments.
}
