% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSnowCourse.R
\name{getSnowCourse}
\alias{getSnowCourse}
\title{Snow course data to database}
\usage{
getSnowCourse(
  hydro_db_path,
  snow_db_path = "//carver/infosys/Snow/DB/SnowDB.mdb",
  inactive = FALSE,
  overwrite = FALSE
)
}
\arguments{
\item{hydro_db_path}{Full path to the hydrology database}

\item{snow_db_path}{Full path to the snow survey database, passed to \code{\link[WRBtools:snowConnect]{WRBtools::snowConnect()}}.}

\item{inactive}{Should inactive stations be included?}

\item{overwrite}{Should all snow course information be overwritten? Only do this if there's been a change in the snow_survey database and you deem this to be the fastest way to reflect that change in the hydro database.}
}
\value{
The hydro database is updated in place.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

\strong{This fuction works with the Microsoft Access snow survey database, circa pre spring 2023}. This function must be refreshed once this database changes.

Pulls and processes snow course data before updating entries into the local hydrometric database.
}
