% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWatersheds.R
\name{getWatersheds}
\alias{getWatersheds}
\title{Add watershed polygons to database}
\usage{
getWatersheds(locations = "WSC", path)
}
\arguments{
\item{locations}{The locations for which you want to add or update polygons. Default "WSC" will get polygons for all WSC stations in the 'locations' table, or specify a vector of location IDs (must all be present in table 'locations'. Currently only supports addition of WSC polygons.}

\item{path}{The path to he database. A new folder will be created at the same level as the database to hold the drainage polygon shapefiles.}
}
\value{
The database table 'polygons' is populated or re-populated with links to shapefiles of polygons, pour points, and station points.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\strong{This function is only capable of fetching pre-made polygons for WSC locations}. Improvements are planned to allow for creation of new polygons using function \code{\link[YGWater:drainageBasins]{YGWater::drainageBasins()}}.

Add watershed polygons to the database, stored in a separate folder alongside the database itself. Currently works only with WSC stations, as uses existing watershed polygons created by the WSC. Each watershed gets a folder containing the polygon, the associated pour point, and the monitoring location itself. In addition, a single shapefile is created containing only the watershed polygons.
}
