% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_update_daily.R
\name{hydro_update_daily}
\alias{hydro_update_daily}
\title{Daily update of hydro database}
\usage{
hydro_update_daily(con = hydrometConnect(silent = TRUE), timeseries_id = "all")
}
\arguments{
\item{con}{A connection to the database, created with \code{\link[DBI:dbConnect]{DBI::dbConnect()}} or using the utility function \code{\link[=hydrometConnect]{hydrometConnect()}}.}

\item{timeseries_id}{The timeseries_ids you wish to have updated, as character or numeric vector. Defaults to "all".}
}
\value{
The database is updated in-place, and diagnostic messages are printed to the console.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function is intended to be run on a daily or near-daily basis to ensure data integrity. Pulls in new data, calculates statistics where necessary, and performs cross-checks on several tables (see details for more information).
}
\details{
Calls several functions in sequence: \code{\link[=getNewRealtime]{getNewRealtime()}} to pull in new data into the measurements_continuous table, \code{\link[=getNewDiscrete]{getNewDiscrete()}} to pull in new data to the measurements_discrete table, \code{\link[=update_hydat]{update_hydat()}} to check for a new HYDAT database version (hydrometric data from the WSC) and incorporate daily means which differ from those already calculated in the dabatabase, update the datums table with any new datums present in HYDAT \code{\link[=update_hydat_datums]{update_hydat_datums()}}, and calculate new statistics if value for last_daily_calculation is < the value for last_new_data in the timeseries database \code{\link[=calculate_stats]{calculate_stats()}}.

Note that this function calls  to update the realtime tables; stations that were added to the table 'timeseries' since the last run are initialized using a separate process.

Any timeseries labelled as 'WRBtools::aq_download()' in the source_fx column in the timeseries table will need your Aquarius username, password, and server URL present in your .Renviron profile, or those three parameters entered in the column source_fx_args: see \code{\link[WRBtools:aq_download]{WRBtools::aq_download()}} for more information.
}
