% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_update_weekly.R
\name{hydro_update_weekly}
\alias{hydro_update_weekly}
\title{Weekly update of hydro database}
\usage{
hydro_update_weekly(
  path,
  WSC_range = Sys.Date() - 577,
  aquarius = TRUE,
  aquarius_range = "unapproved",
  stage = "Stage.Corrected",
  discharge = "Discharge.Master",
  SWE = "SWE.Corrected",
  depth = "Snow Depth.TempCompensated.Corrected",
  distance = "Distance.Corrected",
  server = "https://yukon.aquaticinformatics.net/AQUARIUS"
)
}
\arguments{
\item{path}{The path to the local hydro SQLite database, with extension.}

\item{WSC_range}{The starting date from which to pull real-time WSC data from the web and replace in the local database. Default is max possible days.}

\item{aquarius}{TRUE if you are fetching data from Aquarius, in which case you should also check the next six parameters. FALSE will only populate with WSC data.}

\item{aquarius_range}{Should only unapproved (locked) data be replaced, or all available data? Select from "all" or "unapproved". Default is "unapproved".}

\item{stage}{The name of the stage (level) timeseries as it appears in Aquarius, if it exists, in the form Parameter.Label. All stations must have the same parameter and label. !This DOES NOT apply to WSC stations mirrored in Aquarius.}

\item{discharge}{The name of the discharge (flow) timeseries as it appears in Aquarius, if it exists, in the form Parameter.Label. All stations must have the same parameter and label. !This DOES NOT apply to WSC stations mirrored in Aquarius.}

\item{SWE}{The name of the snow water equivalent timeseries as it appears in Aquarius, if it exists, in the form Parameter.Label. All stations must have the same parameter and label.}

\item{depth}{The name of the snow depth timeseries as it appears in Aquarius, if it exists, in the form Parameter.Label. All stations must have the same parameter and label.}

\item{distance}{The name of the distance timeseries as it appears in Aquarius if it exists, in the form Parameter.Label. All stations must have the same parameter and label. Usually used for distance from bridge girders to water surface.}

\item{server}{The URL to your Aquarius server, if needed. Note that your credentials must be in your .Renviron profile: see ?WRBtools::aq_download.}
}
\value{
Updated entries in the hydro database.
}
\description{
The weekly update function pulls and replaces a large tranche of Water Survey of Canada and Aquarius data to ensure incorporation of any edits. By default, all real-time water survey data currently available online is replaced, as is any Aquarius data not previously labelled as "approved". Daily means and statistics are recalculated for any potentially affected days in the daily tables, except for daily means provided in HYDAT historical tables.
}
