% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_update_weekly.R
\name{hydro_update_weekly}
\alias{hydro_update_weekly}
\title{Weekly update of hydro database}
\usage{
hydro_update_weekly(
  con = hydrometConnect(silent = TRUE),
  timeseries_id = "all",
  start_datetime
)
}
\arguments{
\item{con}{A connection to the database, created with \code{\link[DBI:dbConnect]{DBI::dbConnect()}} or using the utility function \code{\link[=hydrometConnect]{hydrometConnect()}}.}

\item{timeseries_id}{The timeseries_ids you wish to have updated, as character or numeric vector. Defaults to "all".}

\item{start_datetime}{The datetime (as a POSIXct) from which to look for possible new data. You can specify a single start_datetime to apply to all \code{timeseries_id}, or one per element of \code{timeseries_id.}}
}
\value{
Updated entries in the hydro database.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

The weekly update function pulls and replaces a large tranche of Water Survey of Canada and Aquarius data to ensure incorporation of any edits. By default, all real-time water survey data currently available online is replaced, as is any Aquarius data not previously labelled as "approved". Daily means and statistics are recalculated for any potentially affected days in the daily tables, except for daily means provided in HYDAT historical tables.

Any timeseries labelled as 'getRealtimeAquarius' in the source_fx column in the timeseries table will need your Aquarius username, password, and server address present in your .Renviron profile: see \code{\link[=getRealtimeAquarius]{getRealtimeAquarius()}} for more information.
}
