% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeMissing.R
\name{imputeMissing}
\alias{imputeMissing}
\title{Impute missing values}
\usage{
imputeMissing(
  tsid,
  radius,
  start,
  end,
  extra_params = NULL,
  imputed = TRUE,
  daily = FALSE,
  con = hydrometConnect(silent = TRUE)
)
}
\arguments{
\item{tsid}{The target timeseries_id.}

\item{radius}{The radius in kilometers within which to search for similar timeseries.}

\item{start}{The start datetime (as POSIXct object) from which to fill in missing values (can use local or other time zone).}

\item{end}{The end datetime (as POSIXct object) to which to fill in missing values (can use local or other time zone).}

\item{extra_params}{Optional extra parameters to consider when imputing values. For example, you may choose to use wind speeds at 1 or 10 meters to impute speeds at 5 meters.}

\item{imputed}{Should already imputed data be imputed again?}

\item{daily}{Should the imputation be done on a the daily table? Even if set to TRUE this will only apply if there are no entries in table measurements_continuous.}

\item{con}{A connection to the database.}
}
\value{
Imputed values added to the database.
}
\description{
\subsection{Note}{

See function \code{\link[=predictMissing]{predictMissing()}} for machine learning predictions, especially useful if there are no timeseries than be be used with a direct relationship such as predicting flow using level, or level using multiple other level stations.

Impute missing values using either a nearby timeseries or using linear or spline interpolation on the time-series itself. If using nearby timeseries, they can be of same or of other parameters, as specified with parameter 'extra_params'. User interaction is necessary to select the most appropriate timeseries and to review the result of imputation before updating the database.

Imputing long periods of missing data using spline or linear interpolation on the timeseries itself is not recommended. If this is the only feasible option presented by this function because no nearby timeseries can be used for imputation, we recommend using \code{\link[=predictMissing]{predictMissing()}} instead.
}
}
