% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeMissing.R
\name{imputeMissing}
\alias{imputeMissing}
\title{Impute missing values}
\usage{
imputeMissing(
  tsid,
  radius,
  start,
  end,
  extra_params = NULL,
  imputed = TRUE,
  daily = FALSE,
  min_gap = 1,
  max_gap = Inf,
  con = AquaConnect(silent = TRUE)
)
}
\arguments{
\item{tsid}{The target timeseries_id.}

\item{radius}{The radius in kilometers within which to search for similar timeseries.}

\item{start}{The start datetime (as POSIXct, Date, or character) from which to fill in missing values. If specifying a POSIXct the time zone will be respected.}

\item{end}{The end datetime (as POSIXct, Date, or character) to which to fill in missing values. If specifying a POSIXct the time zone will be respected.}

\item{extra_params}{Optional extra parameters (by name, which must match exactly an entry in the database) to consider when imputing values. For example, you may choose to use wind speeds at 1 or 10 meters to impute speeds at 5 meters.}

\item{imputed}{Should already imputed data be imputed again?}

\item{daily}{Should the imputation be done on the daily table? Even if set to TRUE this will only apply if there are no entries in table measurements_continuous to modify.}

\item{min_gap}{An optional integer specifying the minimum number of missing points to interpolate. This can be useful when you want to use a certain method to impute only short gaps, and use another method for longer gaps.}

\item{max_gap}{An optional integer specifying the maximum number of missing points to interpolate. This can be useful when you want to use a certain method to impute only short gaps, and use another method for longer gaps.}

\item{con}{A connection to the database.}
}
\value{
Imputed values added to the database.
}
\description{
\subsection{Note}{

See function predictMissing() (not finished yet!) for machine learning predictions, especially useful if there are no timeseries than be be used with a direct relationship such as predicting flow using level, or level using multiple other level stations.

Impute missing values using either a nearby timeseries or using linear or spline interpolation on the time-series itself. If using nearby timeseries, they can be of same or of other parameters, as specified with parameter 'extra_params'. User interaction is necessary to select the most appropriate timeseries and to review the result of imputation before updating the database.

Imputing long periods of missing data using spline or linear interpolation on the timeseries itself is not recommended. If this is the only feasible option presented by this function because no nearby timeseries can be used for imputation, we recommend using \code{\link[=predictMissing]{predictMissing()}} instead.
}
}
