% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{inf_to_na}
\alias{inf_to_na}
\title{Replace infinite values with NA}
\usage{
inf_to_na(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Numeric vector with infinite values converted to \code{NA}.
}
\description{
Utility function to replace \code{Inf} and \code{-Inf} values with \code{NA}.
}
\keyword{internal}
