% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertACImage.R
\name{insertACImage}
\alias{insertACImage}
\title{Add images to aquacache database}
\usage{
insertACImage(
  object,
  img_meta_id,
  datetime,
  fetch_datetime = NULL,
  description = NULL,
  owner = NULL,
  contributor = NULL,
  share_with = 1,
  location = NULL,
  image_type = NULL,
  con = NULL
)
}
\arguments{
\item{object}{Valid path including extension to the document to upload, or an object of class 'response' such as that provided by \code{\link[=downloadWSCImages]{downloadWSCImages()}}.}

\item{img_meta_id}{The img_meta_id, from the table images_index, corresponding to the image location and type. Set NULL if there is no img_meta_id yet.}

\item{datetime}{The datetime the image was taken at, as a POSIXct object or something that can be coerced to one. If not POSIXct timezone is assumed to be UTC.}

\item{fetch_datetime}{The datetime the image was retrieved (optional). If not POSIXct timezone is assumed to be UTC.}

\item{description}{A description of the image. Pass as text.}

\item{owner}{The owner of the image. Pass as text.}

\item{contributor}{The contributor of the image. Pass as text.}

\item{share_with}{Which user groups should the image be shared with. Default is '1', the public group. Pass as a numeric vector.}

\item{location}{If no img_meta_id exists yet: the location or location_id with which to associate the document (must be in the database). Pass a location code as text and a location_id as a numeric. If img_meta_id is specified, this parameter is ignored.}

\item{image_type}{If no img_meta_id exists yet: the type of image: 'auto', or 'manual'. Pass as text.}

\item{con}{A connection to the database. Default NULL uses AquaConnect() and close the connection afterwards.}
}
\value{
TRUE if an image was properly added to the database.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Images auto-generated on a regular basis should be updated in the database using function \code{\link[=getNewImages]{getNewImages()}} instead of this function. For one-off images, read on.

This function facilitates the addition of one image at a time to the database in the 'images' table. Each image must be linked to a specific location_id from the locations table. Adding an image directly to the database is not possible since the file must be converted to a binary object before loading. See \code{\link[YGwater:getImage]{YGwater::getImage()}} to get the image out again.
}
