% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertHydrometDocument.R
\name{insertHydrometDocument}
\alias{insertHydrometDocument}
\title{Add document to hydrometric database}
\usage{
insertHydrometDocument(
  path,
  locations = NULL,
  lines = NULL,
  polygons = NULL,
  description,
  con = hydrometConnect()
)
}
\arguments{
\item{path}{Valid path including extension to the document to upload.}

\item{locations}{The location(s) with which to associate the document (must be in the database). Leave NULL if associating instead with lines and/or polygons.}

\item{lines}{The line(s) with which to associate the document (must be in the database). Leave NULL if associating instead with locations and/or polygons.}

\item{polygons}{The polygons with which to associate the document (must be in the database). Leave NULL if associating instead with lines and/or locations.}

\item{description}{A text description of what the document is. Please be descriptive!}

\item{con}{A connection to the database, created with \code{\link[DBI:dbConnect]{DBI::dbConnect()}} or using the utility function \code{\link[=hydrometConnect]{hydrometConnect()}}.}
}
\value{
TRUE if a document was properly added to the database.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function facilitates the addition of one document at a time to the database in the 'documents' table. Each document must be linked to a specific location. Adding a document directly to the database is not possible, since the file must be converted to a binary object before loading. See \code{\link[=fetchDocument]{fetchDocument()}} to get a document out again.
\subsection{Locations, lines, and polygons}{

Any document can be associated with locations (points), lines, polygons, or any combination thereof. For example, to associate a document with location '09EA004', line 'Klondike River Near Dempster Bridge', and polygon '09EA004', you would specify locations = '09EA004', lines = 'Klondike River Near Dempster Bridge', polygons = '09EA004'.
}
}
