% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSnowWorkbook.R
\name{readSnowWorkbook}
\alias{readSnowWorkbook}
\title{Read snow workbook and import into Snow database}
\usage{
readSnowWorkbook(workbook, con = snowConnect(silent = TRUE))
}
\arguments{
\item{workbook}{The name of the workbook (excel workbook) containing the snow data.}

\item{con}{A connection to the snow database.}
}
\value{
Does not return any object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Reads snow workbooks created with \code{\link[YGwater:createSnowTemplate]{YGwater::createSnowTemplate()}}, performs some QA/QC, and imports the data to the snow database.
}
