% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSnowWorkbook.R
\name{readSnowWorkbook}
\alias{readSnowWorkbook}
\title{Read snow workbook and import into Snow database}
\usage{
readSnowWorkbook(workbook = "choose", overwrite = FALSE, con = NULL)
}
\arguments{
\item{workbook}{The path to the workbook (.xlsx) containing the snow data. Default "choose" lets you pick the file interactively.}

\item{overwrite}{If \code{TRUE}, will overwrite existing data in the snow database if there's already an entry for the same survey date, target date, and location (regardless of parameters).}

\item{con}{A connection to the snow database. Leave NULL to use function snowConnect() with defaults and close the connection after the function is done.}
}
\value{
Does not return any object. The function is designed to import data into the snow database.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Reads snow workbooks created with \code{\link[YGwater:createSnowTemplate]{YGwater::createSnowTemplate()}}, performs some QA/QC, and imports the data to the snow database. Designed with significant error catching and logging. As the function works through the workbook it may fail on any sheet but will continue to the next sheet until all have been processed. Warning messages will be shown alerting the user explaining the issue and the workbook sheet involved. These warning messages are designed primarily for error catching when this function is run programmatically, but are nevertheless useful for manual use as well.
}
