% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowPop.R
\name{snowPop}
\alias{snowPop}
\title{Initial population of PostgreSQL snow database.}
\usage{
snowPop(
  old_snow_db_path = "//carver/infosys/Snow/DB/SnowDB.mdb",
  con = snowConnect_pg()
)
}
\arguments{
\item{old_snow_db_path}{the path to where the old Access snow database exists}

\item{con}{A connection to the database, created with \code{\link[DBI:dbConnect]{DBI::dbConnect()}} or using the utility function \code{\link[=snowConnect]{snowConnect()}}.}
}
\value{
A populated snowDB database.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Populates the snowDB PostgreSQL database. Pulls existing data from Access database. Currently only populates a couple of tables and is a work in progess.
}
