% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowPop.R
\name{snowPop}
\alias{snowPop}
\title{Initial population of PostgreSQL snow database.}
\usage{
snowPop(
  old_snow_db_path = "//carver/infosys/Snow/DB/SnowDB.mdb",
  con = snowConnect(),
  overwrite = TRUE,
  basins_shp_path =
    "//env-fs/env-data/corp/water/Hydrology/11_SnowMet_Network/02_Manual_Surveys/04_Miscellaneous/Basins_shapefiles/swe_basins.shp"
)
}
\arguments{
\item{old_snow_db_path}{the path to where the old Access snow database exists}

\item{con}{A connection to the database, created with \code{\link[DBI:dbConnect]{DBI::dbConnect()}} or using the utility function \code{\link[=snowConnect]{snowConnect()}}.}

\item{overwrite}{If TRUE, content of tables will be deleted before re-populating. All data in db will be lost!}

\item{basins_shp_path}{Path and file name of sub basins shapefile.}
}
\value{
A populated snowDB database.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Populates the snowDB PostgreSQL database. Pulls existing data from Access database.
}
