% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synchronize_continuous.R
\name{synchronize_continuous}
\alias{synchronize_continuous}
\title{Synchronize hydro DB with remote sources}
\usage{
synchronize_continuous(
  con = NULL,
  timeseries_id = "all",
  start_datetime,
  active = "default",
  sync_remote_false = FALSE,
  dbName = NULL,
  dbHost = NULL,
  dbPort = NULL,
  dbUser = NULL,
  dbPass = NULL
)
}
\arguments{
\item{con}{A connection to the database, created with \code{\link[DBI:dbConnect]{DBI::dbConnect()}} or using the utility function \code{\link[=AquaConnect]{AquaConnect()}}. NULL will create a connection and close it afterwards, otherwise it's up to you to close it after. If you wish to run this function in parallel you MUST leave this argument NULL. If you also specify connection parameters in later arguments they will be used, otherwise the function will use the AquaConnect defaults.}

\item{timeseries_id}{The timeseries_ids you wish to have updated, as character or numeric vector. Defaults to "all".}

\item{start_datetime}{The datetime (as a POSIXct, Date, or character) from which to look for possible new data. You can specify a single start_datetime to apply to all \code{timeseries_id}, or one per element of \code{timeseries_id.}}

\item{active}{Sets behavior for checking timeseries or not. If set to 'default', the function will look to the column 'active' in the 'timeseries' table to determine if new data should be fetched. If set to 'all', the function will ignore the 'active' column and check all timeseries}

\item{sync_remote_false}{Controls whether to synchronize timeseries that have the \code{sync_remote} column set to FALSE in the \code{timeseries} table. Use with extreme caution as setting to TRUE will override the timeseries' intended behavior and overwrite local data with remote data.}

\item{dbName}{The name of the database to connect to. If left NULL, the function will use the default database name from the .Renviron file as per \code{\link[=AquaConnect]{AquaConnect()}}.}

\item{dbHost}{The host address of the database. If left NULL, the function will use the default host address from the .Renviron file as per \code{\link[=AquaConnect]{AquaConnect()}}.}

\item{dbPort}{The port of the database. If left NULL, the function will use the default port from the .Renviron file as per \code{\link[=AquaConnect]{AquaConnect()}}.}

\item{dbUser}{The username for the database. If left NULL, the function will use the default username from the .Renviron file as per \code{\link[=AquaConnect]{AquaConnect()}}.}

\item{dbPass}{The password for the database. If left NULL, the function will use the default password from the .Renviron file as per \code{\link[=AquaConnect]{AquaConnect()}}.}
}
\value{
A data.frame showing the status of synchronization for each timeseries as well as the error or warning message, if any, plus updated entries in the hydro database.
}
\description{
This synchronize function pulls and replaces data referenced in table 'timeseries' if and when a discrepancy is observed between the remote repository and the local data store, with the remote taking precedence. New data is also brought in, if any exists on the remote. Daily means and statistics are recalculated for any potentially affected days in the daily tables, except for daily means provided in HYDAT historical tables (Water Survey of Canada).

If you leave the 'con' object as NULL the function will try to run in parallel, which dramatically speeds things up. You can pass connection parameters to the function if you want to run it in parallel, but you must leave 'con' as NULL. If you don't want to run in parallel, you can pass a connection object to the function and it will run in sequence.

In addition, grades, qualifiers, and approvals are always updated as it's computationally cheaper to do so than to check if they need updating.

NOTE that any data point labelled as imputed = TRUE is only replaced if a value is found in the remote exactly matching the datetime of the imputed entry, and any data point labelled as no_update = TRUE is not replaced by the remote dat (imputed or not).

A timeseries whose selected synchronization assignment uses
\code{downloadAquarius} needs credentials in \code{.Renviron} or in that assignment's
arguments; see \code{\link[=downloadAquarius]{downloadAquarius()}}.
}
