% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_hydat.R
\name{update_hydat}
\alias{update_hydat}
\title{Update HYDAT-related timeseries}
\usage{
update_hydat(timeseries, path, force_update = FALSE)
}
\arguments{
\item{timeseries}{Character vector of timeseries for which to look for updates. "all" will attempt to update all timeseries from operator 'WSC'.}

\item{path}{The path to the hydrometric database, passed to \code{\link[WRBtools:hydroConnect]{WRBtools::hydroConnect()}}.}

\item{force_update}{Set TRUE if you want to force a check of each location against the local copy of HYDAT.}
}
\value{
Updated daily means where HYDAT values exist, and a boolean indicating if the HYDAT database was in fact updated. Note: if force_update is TRUE but there is no new HYDAT database, FALSE will be returned even though database timeseries are updated.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

First checks the local version of HYDAT using function \code{\link[WRBtools:hydat_check]{WRBtools::hydat_check()}} and updates if needed, then checks the local copy against the one last used by the database. If needed or if force_update == TRUE, proceeds to checking each location specified for new data and replaced old data wherever a discrepancy is noted. If all WSC timeseries in the WRB hydro database are in timeseries, will also update the internal_status table with the HYDAT version used for the update.
}
