% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_hydat.R
\name{update_hydat}
\alias{update_hydat}
\title{Update HYDAT-related timeseries}
\usage{
update_hydat(
  con = hydrometConnect(silent = TRUE),
  timeseries_id = "all",
  force_update = FALSE
)
}
\arguments{
\item{con}{A connection to the database, created with \code{\link[DBI:dbConnect]{DBI::dbConnect()}} or using the utility function \code{\link[=hydrometConnect]{hydrometConnect()}}.}

\item{timeseries_id}{Character vector of timeseries_ids for which to look for updates. "all" will attempt to update all timeseries from operator 'WSC'.}

\item{force_update}{Set TRUE if you want to force a check of each location against the local copy of HYDAT.}
}
\value{
Updated daily means where HYDAT values were replaced by updated values, and TRUE if a new version of hydat was found.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

First checks the local version of HYDAT using function \code{\link[WRBtools:hydat_check]{WRBtools::hydat_check()}} and updates if needed, then checks the local copy against the one last used by the database. If needed or if force_update == TRUE, proceeds to checking each location specified for new data and replaced old data wherever a discrepancy is noted. If all WSC timeseries in the WRB hydro database are in timeseries, will also update the internal_status table with the HYDAT version used for the update.
}
